Implement new ref browser based on bootstrap 5.2 + vuejs + restapi.
