--- ftw/referencewidget/resources/main.css
@@ -1,1 +1,0 @@
-@import url('refwidget.css');
