--- ftw/referencewidget/resources/main.js
@@ -1,1 +1,0 @@
-require(['handlebars', 'refwidget', 'refwidget-tinymce'], function (handlebars, refwidget, refwidgetTinymce) {});
