--- ftw/referencewidget/resources/referencewidget-compiled.js
@@ -1,35 +1,0 @@
-/*!
-
- handlebars v4.0.5
-
-Copyright (C) 2011-2015 by Yehuda Katz
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
-
-@license
-*/
-
-/**
- * @license text 2.0.16 Copyright jQuery Foundation and other contributors.
- * Released under MIT license, http://github.com/requirejs/text/LICENSE
- */
-
-!function(t,e){"object"==typeof exports&&"object"==typeof module?module.exports=e():"object"==typeof exports?exports.Handlebars=e():t.Handlebars=e()}(this,function(){return function(t){function e(n){if(r[n])return r[n].exports;var i=r[n]={exports:{},id:n,loaded:!1};return t[n].call(i.exports,i,i.exports,e),i.loaded=!0,i.exports}var r={};return e.m=t,e.c=r,e.p="",e(0)}([function(t,e,r){"use strict";function n(){var t=v();return t.compile=function(e,r){return p.compile(e,r,t)},t.precompile=function(e,r){return p.precompile(e,r,t)},t.AST=l.default,t.Compiler=p.Compiler,t.JavaScriptCompiler=u.default,t.Parser=c.parser,t.parse=c.parse,t}var i=r(1).default;e.__esModule=!0;var a=r(2),s=i(a),o=r(21),l=i(o),c=r(22),p=r(27),h=r(28),u=i(h),d=r(25),f=i(d),m=r(20),g=i(m),v=s.default.create,y=n();y.create=n,g.default(y),y.Visitor=f.default,y.default=y,e.default=y,t.exports=e.default},function(t,e){"use strict";e.default=function(t){return t&&t.__esModule?t:{default:t}},e.__esModule=!0},function(t,e,r){"use strict";function n(){var t=new o.HandlebarsEnvironment;return d.extend(t,o),t.SafeString=c.default,t.Exception=h.default,t.Utils=d,t.escapeExpression=d.escapeExpression,t.VM=m,t.template=function(e){return m.template(e,t)},t}var i=r(3).default,a=r(1).default;e.__esModule=!0;var s=r(4),o=i(s),l=r(18),c=a(l),p=r(6),h=a(p),u=r(5),d=i(u),f=r(19),m=i(f),g=r(20),v=a(g),y=n();y.create=n,v.default(y),y.default=y,e.default=y,t.exports=e.default},function(t,e){"use strict";e.default=function(t){if(t&&t.__esModule)return t;var e={};if(null!=t)for(var r in t)Object.prototype.hasOwnProperty.call(t,r)&&(e[r]=t[r]);return e.default=t,e},e.__esModule=!0},function(t,e,r){"use strict";function n(t,e,r){this.helpers=t||{},this.partials=e||{},this.decorators=r||{},l.registerDefaultHelpers(this),c.registerDefaultDecorators(this)}var i=r(1).default;e.__esModule=!0,e.HandlebarsEnvironment=n;var a=r(5),s=r(6),o=i(s),l=r(7),c=r(15),p=r(17),h=i(p);e.VERSION="4.0.5";e.COMPILER_REVISION=7;var u={1:"<= 1.0.rc.2",2:"== 1.0.0-rc.3",3:"== 1.0.0-rc.4",4:"== 1.x.x",5:"== 2.0.0-alpha.x",6:">= 2.0.0-beta.1",7:">= 4.0.0"};e.REVISION_CHANGES=u;n.prototype={constructor:n,logger:h.default,log:h.default.log,registerHelper:function(t,e){if("[object Object]"===a.toString.call(t)){if(e)throw new o.default("Arg not supported with multiple helpers");a.extend(this.helpers,t)}else this.helpers[t]=e},unregisterHelper:function(t){delete this.helpers[t]},registerPartial:function(t,e){if("[object Object]"===a.toString.call(t))a.extend(this.partials,t);else{if(void 0===e)throw new o.default('Attempting to register a partial called "'+t+'" as undefined');this.partials[t]=e}},unregisterPartial:function(t){delete this.partials[t]},registerDecorator:function(t,e){if("[object Object]"===a.toString.call(t)){if(e)throw new o.default("Arg not supported with multiple decorators");a.extend(this.decorators,t)}else this.decorators[t]=e},unregisterDecorator:function(t){delete this.decorators[t]}};var d=h.default.log;e.log=d,e.createFrame=a.createFrame,e.logger=h.default},function(t,e){"use strict";function r(t){return p[t]}function n(t){for(var e=1;e<arguments.length;e++)for(var r in arguments[e])Object.prototype.hasOwnProperty.call(arguments[e],r)&&(t[r]=arguments[e][r]);return t}function i(t,e){for(var r=0,n=t.length;r<n;r++)if(t[r]===e)return r;return-1}function a(t){if("string"!=typeof t){if(t&&t.toHTML)return t.toHTML();if(null==t)return"";if(!t)return t+"";t=""+t}return u.test(t)?t.replace(h,r):t}function s(t){return!t&&0!==t||!(!m(t)||0!==t.length)}function o(t){var e=n({},t);return e._parent=t,e}function l(t,e){return t.path=e,t}function c(t,e){return(t?t+".":"")+e}e.__esModule=!0,e.extend=n,e.indexOf=i,e.escapeExpression=a,e.isEmpty=s,e.createFrame=o,e.blockParams=l,e.appendContextPath=c;var p={"&":"&amp;","<":"&lt;",">":"&gt;",'"':"&quot;","'":"&#x27;","`":"&#x60;","=":"&#x3D;"},h=/[&<>"'`=]/g,u=/[&<>"'`=]/,d=Object.prototype.toString;e.toString=d;var f=function(t){return"function"==typeof t};f(/x/)&&(e.isFunction=f=function(t){return"function"==typeof t&&"[object Function]"===d.call(t)}),e.isFunction=f;var m=Array.isArray||function(t){return!(!t||"object"!=typeof t)&&"[object Array]"===d.call(t)};e.isArray=m},function(t,e){"use strict";function r(t,e){var i=e&&e.loc,a=void 0,s=void 0;i&&(a=i.start.line,s=i.start.column,t+=" - "+a+":"+s);for(var o=Error.prototype.constructor.call(this,t),l=0;l<n.length;l++)this[n[l]]=o[n[l]];Error.captureStackTrace&&Error.captureStackTrace(this,r),i&&(this.lineNumber=a,this.column=s)}e.__esModule=!0;var n=["description","fileName","lineNumber","message","name","number","stack"];r.prototype=new Error,e.default=r,t.exports=e.default},function(t,e,r){"use strict";function n(t){s.default(t),l.default(t),p.default(t),u.default(t),f.default(t),g.default(t),y.default(t)}var i=r(1).default;e.__esModule=!0,e.registerDefaultHelpers=n;var a=r(8),s=i(a),o=r(9),l=i(o),c=r(10),p=i(c),h=r(11),u=i(h),d=r(12),f=i(d),m=r(13),g=i(m),v=r(14),y=i(v)},function(t,e,r){"use strict";e.__esModule=!0;var n=r(5);e.default=function(t){t.registerHelper("blockHelperMissing",function(e,r){var i=r.inverse,a=r.fn;if(!0===e)return a(this);if(!1===e||null==e)return i(this);if(n.isArray(e))return e.length>0?(r.ids&&(r.ids=[r.name]),t.helpers.each(e,r)):i(this);if(r.data&&r.ids){var s=n.createFrame(r.data);s.contextPath=n.appendContextPath(r.data.contextPath,r.name),r={data:s}}return a(e,r)})},t.exports=e.default},function(t,e,r){"use strict";var n=r(1).default;e.__esModule=!0;var i=r(5),a=r(6),s=n(a);e.default=function(t){t.registerHelper("each",function(t,e){function r(e,r,a){c&&(c.key=e,c.index=r,c.first=0===r,c.last=!!a,p&&(c.contextPath=p+e)),l+=n(t[e],{data:c,blockParams:i.blockParams([t[e],e],[p+e,null])})}if(!e)throw new s.default("Must pass iterator to #each");var n=e.fn,a=e.inverse,o=0,l="",c=void 0,p=void 0;if(e.data&&e.ids&&(p=i.appendContextPath(e.data.contextPath,e.ids[0])+"."),i.isFunction(t)&&(t=t.call(this)),e.data&&(c=i.createFrame(e.data)),t&&"object"==typeof t)if(i.isArray(t))for(var h=t.length;o<h;o++)o in t&&r(o,o,o===t.length-1);else{var u=void 0;for(var d in t)t.hasOwnProperty(d)&&(void 0!==u&&r(u,o-1),u=d,o++);void 0!==u&&r(u,o-1,!0)}return 0===o&&(l=a(this)),l})},t.exports=e.default},function(t,e,r){"use strict";var n=r(1).default;e.__esModule=!0;var i=r(6),a=n(i);e.default=function(t){t.registerHelper("helperMissing",function(){if(1!==arguments.length)throw new a.default('Missing helper: "'+arguments[arguments.length-1].name+'"')})},t.exports=e.default},function(t,e,r){"use strict";e.__esModule=!0;var n=r(5);e.default=function(t){t.registerHelper("if",function(t,e){return n.isFunction(t)&&(t=t.call(this)),!e.hash.includeZero&&!t||n.isEmpty(t)?e.inverse(this):e.fn(this)}),t.registerHelper("unless",function(e,r){return t.helpers.if.call(this,e,{fn:r.inverse,inverse:r.fn,hash:r.hash})})},t.exports=e.default},function(t,e){"use strict";e.__esModule=!0,e.default=function(t){t.registerHelper("log",function(){for(var e=[void 0],r=arguments[arguments.length-1],n=0;n<arguments.length-1;n++)e.push(arguments[n]);var i=1;null!=r.hash.level?i=r.hash.level:r.data&&null!=r.data.level&&(i=r.data.level),e[0]=i,t.log.apply(t,e)})},t.exports=e.default},function(t,e){"use strict";e.__esModule=!0,e.default=function(t){t.registerHelper("lookup",function(t,e){return t&&t[e]})},t.exports=e.default},function(t,e,r){"use strict";e.__esModule=!0;var n=r(5);e.default=function(t){t.registerHelper("with",function(t,e){n.isFunction(t)&&(t=t.call(this));var r=e.fn;if(n.isEmpty(t))return e.inverse(this);var i=e.data;return e.data&&e.ids&&(i=n.createFrame(e.data),i.contextPath=n.appendContextPath(e.data.contextPath,e.ids[0])),r(t,{data:i,blockParams:n.blockParams([t],[i&&i.contextPath])})})},t.exports=e.default},function(t,e,r){"use strict";function n(t){s.default(t)}var i=r(1).default;e.__esModule=!0,e.registerDefaultDecorators=n;var a=r(16),s=i(a)},function(t,e,r){"use strict";e.__esModule=!0;var n=r(5);e.default=function(t){t.registerDecorator("inline",function(t,e,r,i){var a=t;return e.partials||(e.partials={},a=function(i,a){var s=r.partials;r.partials=n.extend({},s,e.partials);var o=t(i,a);return r.partials=s,o}),e.partials[i.args[0]]=i.fn,a})},t.exports=e.default},function(t,e,r){"use strict";e.__esModule=!0;var n=r(5),i={methodMap:["debug","info","warn","error"],level:"info",lookupLevel:function(t){if("string"==typeof t){var e=n.indexOf(i.methodMap,t.toLowerCase());t=e>=0?e:parseInt(t,10)}return t},log:function(t){if(t=i.lookupLevel(t),"undefined"!=typeof console&&i.lookupLevel(i.level)<=t){var e=i.methodMap[t];console[e]||(e="log");for(var r=arguments.length,n=Array(r>1?r-1:0),a=1;a<r;a++)n[a-1]=arguments[a];console[e].apply(console,n)}}};e.default=i,t.exports=e.default},function(t,e){"use strict";function r(t){this.string=t}e.__esModule=!0,r.prototype.toString=r.prototype.toHTML=function(){return""+this.string},e.default=r,t.exports=e.default},function(t,e,r){"use strict";function n(t){var e=t&&t[0]||1,r=v.COMPILER_REVISION;if(e!==r){if(e<r){var n=v.REVISION_CHANGES[r],i=v.REVISION_CHANGES[e];throw new g.default("Template was precompiled with an older version of Handlebars than the current runtime. Please update your precompiler to a newer version ("+n+") or downgrade your runtime to an older version ("+i+").")}throw new g.default("Template was precompiled with a newer version of Handlebars than the current runtime. Please update your runtime to a newer version ("+t[1]+").")}}function i(t,e){function r(r,n,i){i.hash&&(n=f.extend({},n,i.hash),i.ids&&(i.ids[0]=!0)),r=e.VM.resolvePartial.call(this,r,n,i);var a=e.VM.invokePartial.call(this,r,n,i);if(null==a&&e.compile&&(i.partials[i.name]=e.compile(r,t.compilerOptions,e),a=i.partials[i.name](n,i)),null!=a){if(i.indent){for(var s=a.split("\n"),o=0,l=s.length;o<l&&(s[o]||o+1!==l);o++)s[o]=i.indent+s[o];a=s.join("\n")}return a}throw new g.default("The partial "+i.name+" could not be compiled when running in runtime-only mode")}function n(e){function r(e){return""+t.main(i,e,i.helpers,i.partials,s,l,o)}var a=arguments.length<=1||void 0===arguments[1]?{}:arguments[1],s=a.data;n._setup(a),!a.partial&&t.useData&&(s=c(e,s));var o=void 0,l=t.useBlockParams?[]:void 0;return t.useDepths&&(o=a.depths?e!==a.depths[0]?[e].concat(a.depths):a.depths:[e]),(r=p(t.main,r,i,a.depths||[],s,l))(e,a)}if(!e)throw new g.default("No environment passed to template");if(!t||!t.main)throw new g.default("Unknown template object: "+typeof t);t.main.decorator=t.main_d,e.VM.checkRevision(t.compiler);var i={strict:function(t,e){if(!(e in t))throw new g.default('"'+e+'" not defined in '+t);return t[e]},lookup:function(t,e){for(var r=t.length,n=0;n<r;n++)if(t[n]&&null!=t[n][e])return t[n][e]},lambda:function(t,e){return"function"==typeof t?t.call(e):t},escapeExpression:f.escapeExpression,invokePartial:r,fn:function(e){var r=t[e];return r.decorator=t[e+"_d"],r},programs:[],program:function(t,e,r,n,i){var s=this.programs[t],o=this.fn(t);return e||i||n||r?s=a(this,t,o,e,r,n,i):s||(s=this.programs[t]=a(this,t,o)),s},data:function(t,e){for(;t&&e--;)t=t._parent;return t},merge:function(t,e){var r=t||e;return t&&e&&t!==e&&(r=f.extend({},e,t)),r},noop:e.VM.noop,compilerInfo:t.compiler};return n.isTop=!0,n._setup=function(r){r.partial?(i.helpers=r.helpers,i.partials=r.partials,i.decorators=r.decorators):(i.helpers=i.merge(r.helpers,e.helpers),t.usePartial&&(i.partials=i.merge(r.partials,e.partials)),(t.usePartial||t.useDecorators)&&(i.decorators=i.merge(r.decorators,e.decorators)))},n._child=function(e,r,n,s){if(t.useBlockParams&&!n)throw new g.default("must pass block params");if(t.useDepths&&!s)throw new g.default("must pass parent depths");return a(i,e,t[e],r,0,n,s)},n}function a(t,e,r,n,i,a,s){function o(e){var i=arguments.length<=1||void 0===arguments[1]?{}:arguments[1],o=s;return s&&e!==s[0]&&(o=[e].concat(s)),r(t,e,t.helpers,t.partials,i.data||n,a&&[i.blockParams].concat(a),o)}return o=p(r,o,t,s,n,a),o.program=e,o.depth=s?s.length:0,o.blockParams=i||0,o}function s(t,e,r){return t?t.call||r.name||(r.name=t,t=r.partials[t]):t="@partial-block"===r.name?r.data["partial-block"]:r.partials[r.name],t}function o(t,e,r){r.partial=!0,r.ids&&(r.data.contextPath=r.ids[0]||r.data.contextPath);var n=void 0;if(r.fn&&r.fn!==l&&(r.data=v.createFrame(r.data),n=r.data["partial-block"]=r.fn,n.partials&&(r.partials=f.extend({},r.partials,n.partials))),void 0===t&&n&&(t=n),void 0===t)throw new g.default("The partial "+r.name+" could not be found");if(t instanceof Function)return t(e,r)}function l(){return""}function c(t,e){return e&&"root"in e||(e=e?v.createFrame(e):{},e.root=t),e}function p(t,e,r,n,i,a){if(t.decorator){var s={};e=t.decorator(e,s,r,n&&n[0],i,a,n),f.extend(e,s)}return e}var h=r(3).default,u=r(1).default;e.__esModule=!0,e.checkRevision=n,e.template=i,e.wrapProgram=a,e.resolvePartial=s,e.invokePartial=o,e.noop=l;var d=r(5),f=h(d),m=r(6),g=u(m),v=r(4)},function(t,e){(function(r){"use strict";e.__esModule=!0,e.default=function(t){var e=void 0!==r?r:window,n=e.Handlebars;t.noConflict=function(){return e.Handlebars===t&&(e.Handlebars=n),t}},t.exports=e.default}).call(e,function(){return this}())},function(t,e){"use strict";e.__esModule=!0;var r={helpers:{helperExpression:function(t){return"SubExpression"===t.type||("MustacheStatement"===t.type||"BlockStatement"===t.type)&&!!(t.params&&t.params.length||t.hash)},scopedId:function(t){return/^\.|this\b/.test(t.original)},simpleId:function(t){return 1===t.parts.length&&!r.helpers.scopedId(t)&&!t.depth}}};e.default=r,t.exports=e.default},function(t,e,r){"use strict";function n(t,e){return"Program"===t.type?t:(o.default.yy=d,d.locInfo=function(t){return new d.SourceLocation(e&&e.srcName,t)},new c.default(e).accept(o.default.parse(t)))}var i=r(1).default,a=r(3).default;e.__esModule=!0,e.parse=n;var s=r(23),o=i(s),l=r(24),c=i(l),p=r(26),h=a(p),u=r(5);e.parser=o.default;var d={};u.extend(d,h)},function(t,e){"use strict";var r=function(){function t(){this.yy={}}var e={trace:function(){},yy:{},symbols_:{error:2,root:3,program:4,EOF:5,program_repetition0:6,statement:7,mustache:8,block:9,rawBlock:10,partial:11,partialBlock:12,content:13,COMMENT:14,CONTENT:15,openRawBlock:16,rawBlock_repetition_plus0:17,END_RAW_BLOCK:18,OPEN_RAW_BLOCK:19,helperName:20,openRawBlock_repetition0:21,openRawBlock_option0:22,CLOSE_RAW_BLOCK:23,openBlock:24,block_option0:25,closeBlock:26,openInverse:27,block_option1:28,OPEN_BLOCK:29,openBlock_repetition0:30,openBlock_option0:31,openBlock_option1:32,CLOSE:33,OPEN_INVERSE:34,openInverse_repetition0:35,openInverse_option0:36,openInverse_option1:37,openInverseChain:38,OPEN_INVERSE_CHAIN:39,openInverseChain_repetition0:40,openInverseChain_option0:41,openInverseChain_option1:42,inverseAndProgram:43,INVERSE:44,inverseChain:45,inverseChain_option0:46,OPEN_ENDBLOCK:47,OPEN:48,mustache_repetition0:49,mustache_option0:50,OPEN_UNESCAPED:51,mustache_repetition1:52,mustache_option1:53,CLOSE_UNESCAPED:54,OPEN_PARTIAL:55,partialName:56,partial_repetition0:57,partial_option0:58,openPartialBlock:59,OPEN_PARTIAL_BLOCK:60,openPartialBlock_repetition0:61,openPartialBlock_option0:62,param:63,sexpr:64,OPEN_SEXPR:65,sexpr_repetition0:66,sexpr_option0:67,CLOSE_SEXPR:68,hash:69,hash_repetition_plus0:70,hashSegment:71,ID:72,EQUALS:73,blockParams:74,OPEN_BLOCK_PARAMS:75,blockParams_repetition_plus0:76,CLOSE_BLOCK_PARAMS:77,path:78,dataName:79,STRING:80,NUMBER:81,BOOLEAN:82,UNDEFINED:83,NULL:84,DATA:85,pathSegments:86,SEP:87,$accept:0,$end:1},terminals_:{2:"error",5:"EOF",14:"COMMENT",15:"CONTENT",18:"END_RAW_BLOCK",19:"OPEN_RAW_BLOCK",23:"CLOSE_RAW_BLOCK",29:"OPEN_BLOCK",33:"CLOSE",34:"OPEN_INVERSE",39:"OPEN_INVERSE_CHAIN",44:"INVERSE",47:"OPEN_ENDBLOCK",48:"OPEN",51:"OPEN_UNESCAPED",54:"CLOSE_UNESCAPED",55:"OPEN_PARTIAL",60:"OPEN_PARTIAL_BLOCK",65:"OPEN_SEXPR",68:"CLOSE_SEXPR",72:"ID",73:"EQUALS",75:"OPEN_BLOCK_PARAMS",77:"CLOSE_BLOCK_PARAMS",80:"STRING",81:"NUMBER",82:"BOOLEAN",83:"UNDEFINED",84:"NULL",85:"DATA",87:"SEP"},productions_:[0,[3,2],[4,1],[7,1],[7,1],[7,1],[7,1],[7,1],[7,1],[7,1],[13,1],[10,3],[16,5],[9,4],[9,4],[24,6],[27,6],[38,6],[43,2],[45,3],[45,1],[26,3],[8,5],[8,5],[11,5],[12,3],[59,5],[63,1],[63,1],[64,5],[69,1],[71,3],[74,3],[20,1],[20,1],[20,1],[20,1],[20,1],[20,1],[20,1],[56,1],[56,1],[79,2],[78,1],[86,3],[86,1],[6,0],[6,2],[17,1],[17,2],[21,0],[21,2],[22,0],[22,1],[25,0],[25,1],[28,0],[28,1],[30,0],[30,2],[31,0],[31,1],[32,0],[32,1],[35,0],[35,2],[36,0],[36,1],[37,0],[37,1],[40,0],[40,2],[41,0],[41,1],[42,0],[42,1],[46,0],[46,1],[49,0],[49,2],[50,0],[50,1],[52,0],[52,2],[53,0],[53,1],[57,0],[57,2],[58,0],[58,1],[61,0],[61,2],[62,0],[62,1],[66,0],[66,2],[67,0],[67,1],[70,1],[70,2],[76,1],[76,2]],performAction:function(t,e,r,n,i,a,s){var o=a.length-1;switch(i){case 1:return a[o-1];case 2:this.$=n.prepareProgram(a[o]);break;case 3:case 4:case 5:case 6:case 7:case 8:this.$=a[o];break;case 9:this.$={type:"CommentStatement",value:n.stripComment(a[o]),strip:n.stripFlags(a[o],a[o]),loc:n.locInfo(this._$)};break;case 10:this.$={type:"ContentStatement",original:a[o],value:a[o],loc:n.locInfo(this._$)};break;case 11:this.$=n.prepareRawBlock(a[o-2],a[o-1],a[o],this._$);break;case 12:this.$={path:a[o-3],params:a[o-2],hash:a[o-1]};break;case 13:this.$=n.prepareBlock(a[o-3],a[o-2],a[o-1],a[o],!1,this._$);break;case 14:this.$=n.prepareBlock(a[o-3],a[o-2],a[o-1],a[o],!0,this._$);break;case 15:this.$={open:a[o-5],path:a[o-4],params:a[o-3],hash:a[o-2],blockParams:a[o-1],strip:n.stripFlags(a[o-5],a[o])};break;case 16:case 17:this.$={path:a[o-4],params:a[o-3],hash:a[o-2],blockParams:a[o-1],strip:n.stripFlags(a[o-5],a[o])};break;case 18:this.$={strip:n.stripFlags(a[o-1],a[o-1]),program:a[o]};break;case 19:var l=n.prepareBlock(a[o-2],a[o-1],a[o],a[o],!1,this._$),c=n.prepareProgram([l],a[o-1].loc);c.chained=!0,this.$={strip:a[o-2].strip,program:c,chain:!0};break;case 20:this.$=a[o];break;case 21:this.$={path:a[o-1],strip:n.stripFlags(a[o-2],a[o])};break;case 22:case 23:this.$=n.prepareMustache(a[o-3],a[o-2],a[o-1],a[o-4],n.stripFlags(a[o-4],a[o]),this._$);break;case 24:this.$={type:"PartialStatement",name:a[o-3],params:a[o-2],hash:a[o-1],indent:"",strip:n.stripFlags(a[o-4],a[o]),loc:n.locInfo(this._$)};break;case 25:this.$=n.preparePartialBlock(a[o-2],a[o-1],a[o],this._$);break;case 26:this.$={path:a[o-3],params:a[o-2],hash:a[o-1],strip:n.stripFlags(a[o-4],a[o])};break;case 27:case 28:this.$=a[o];break;case 29:this.$={type:"SubExpression",path:a[o-3],params:a[o-2],hash:a[o-1],loc:n.locInfo(this._$)};break;case 30:this.$={type:"Hash",pairs:a[o],loc:n.locInfo(this._$)};break;case 31:this.$={type:"HashPair",key:n.id(a[o-2]),value:a[o],loc:n.locInfo(this._$)};break;case 32:this.$=n.id(a[o-1]);break;case 33:case 34:this.$=a[o];break;case 35:this.$={type:"StringLiteral",value:a[o],original:a[o],loc:n.locInfo(this._$)};break;case 36:this.$={type:"NumberLiteral",value:Number(a[o]),original:Number(a[o]),loc:n.locInfo(this._$)};break;case 37:this.$={type:"BooleanLiteral",value:"true"===a[o],original:"true"===a[o],loc:n.locInfo(this._$)};break;case 38:this.$={type:"UndefinedLiteral",original:void 0,value:void 0,loc:n.locInfo(this._$)};break;case 39:this.$={type:"NullLiteral",original:null,value:null,loc:n.locInfo(this._$)};break;case 40:case 41:this.$=a[o];break;case 42:this.$=n.preparePath(!0,a[o],this._$);break;case 43:this.$=n.preparePath(!1,a[o],this._$);break;case 44:a[o-2].push({part:n.id(a[o]),original:a[o],separator:a[o-1]}),this.$=a[o-2];break;case 45:this.$=[{part:n.id(a[o]),original:a[o]}];break;case 46:this.$=[];break;case 47:a[o-1].push(a[o]);break;case 48:this.$=[a[o]];break;case 49:a[o-1].push(a[o]);break;case 50:this.$=[];break;case 51:a[o-1].push(a[o]);break;case 58:this.$=[];break;case 59:a[o-1].push(a[o]);break;case 64:this.$=[];break;case 65:a[o-1].push(a[o]);break;case 70:this.$=[];break;case 71:a[o-1].push(a[o]);break;case 78:this.$=[];break;case 79:a[o-1].push(a[o]);break;case 82:this.$=[];break;case 83:a[o-1].push(a[o]);break;case 86:this.$=[];break;case 87:a[o-1].push(a[o]);break;case 90:this.$=[];break;case 91:a[o-1].push(a[o]);break;case 94:this.$=[];break;case 95:a[o-1].push(a[o]);break;case 98:this.$=[a[o]];break;case 99:a[o-1].push(a[o]);break;case 100:this.$=[a[o]];break;case 101:a[o-1].push(a[o])}},table:[{3:1,4:2,5:[2,46],6:3,14:[2,46],15:[2,46],19:[2,46],29:[2,46],34:[2,46],48:[2,46],51:[2,46],55:[2,46],60:[2,46]},{1:[3]},{5:[1,4]},{5:[2,2],7:5,8:6,9:7,10:8,11:9,12:10,13:11,14:[1,12],15:[1,20],16:17,19:[1,23],24:15,27:16,29:[1,21],34:[1,22],39:[2,2],44:[2,2],47:[2,2],48:[1,13],51:[1,14],55:[1,18],59:19,60:[1,24]},{1:[2,1]},{5:[2,47],14:[2,47],15:[2,47],19:[2,47],29:[2,47],34:[2,47],39:[2,47],44:[2,47],47:[2,47],48:[2,47],51:[2,47],55:[2,47],60:[2,47]},{5:[2,3],14:[2,3],15:[2,3],19:[2,3],29:[2,3],34:[2,3],39:[2,3],44:[2,3],47:[2,3],48:[2,3],51:[2,3],55:[2,3],60:[2,3]},{5:[2,4],14:[2,4],15:[2,4],19:[2,4],29:[2,4],34:[2,4],39:[2,4],44:[2,4],47:[2,4],48:[2,4],51:[2,4],55:[2,4],60:[2,4]},{5:[2,5],14:[2,5],15:[2,5],19:[2,5],29:[2,5],34:[2,5],39:[2,5],44:[2,5],47:[2,5],48:[2,5],51:[2,5],55:[2,5],60:[2,5]},{5:[2,6],14:[2,6],15:[2,6],19:[2,6],29:[2,6],34:[2,6],39:[2,6],44:[2,6],47:[2,6],48:[2,6],51:[2,6],55:[2,6],60:[2,6]},{5:[2,7],14:[2,7],15:[2,7],19:[2,7],29:[2,7],34:[2,7],39:[2,7],44:[2,7],47:[2,7],48:[2,7],51:[2,7],55:[2,7],60:[2,7]},{5:[2,8],14:[2,8],15:[2,8],19:[2,8],29:[2,8],34:[2,8],39:[2,8],44:[2,8],47:[2,8],48:[2,8],51:[2,8],55:[2,8],60:[2,8]},{5:[2,9],14:[2,9],15:[2,9],19:[2,9],29:[2,9],34:[2,9],39:[2,9],44:[2,9],47:[2,9],48:[2,9],51:[2,9],55:[2,9],60:[2,9]},{20:25,72:[1,35],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{20:36,72:[1,35],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{4:37,6:3,14:[2,46],15:[2,46],19:[2,46],29:[2,46],34:[2,46],39:[2,46],44:[2,46],47:[2,46],48:[2,46],51:[2,46],55:[2,46],60:[2,46]},{4:38,6:3,14:[2,46],15:[2,46],19:[2,46],29:[2,46],34:[2,46],44:[2,46],47:[2,46],48:[2,46],51:[2,46],55:[2,46],60:[2,46]},{13:40,15:[1,20],17:39},{20:42,56:41,64:43,65:[1,44],72:[1,35],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{4:45,6:3,14:[2,46],15:[2,46],19:[2,46],29:[2,46],34:[2,46],47:[2,46],48:[2,46],51:[2,46],55:[2,46],60:[2,46]},{5:[2,10],14:[2,10],15:[2,10],18:[2,10],19:[2,10],29:[2,10],34:[2,10],39:[2,10],44:[2,10],47:[2,10],48:[2,10],51:[2,10],55:[2,10],60:[2,10]},{20:46,72:[1,35],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{20:47,72:[1,35],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{20:48,72:[1,35],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{20:42,56:49,64:43,65:[1,44],72:[1,35],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{33:[2,78],49:50,65:[2,78],72:[2,78],80:[2,78],81:[2,78],82:[2,78],83:[2,78],84:[2,78],85:[2,78]},{23:[2,33],33:[2,33],54:[2,33],65:[2,33],68:[2,33],72:[2,33],75:[2,33],80:[2,33],81:[2,33],82:[2,33],83:[2,33],84:[2,33],85:[2,33]},{23:[2,34],33:[2,34],54:[2,34],65:[2,34],68:[2,34],72:[2,34],75:[2,34],80:[2,34],81:[2,34],82:[2,34],83:[2,34],84:[2,34],85:[2,34]},{23:[2,35],33:[2,35],54:[2,35],65:[2,35],68:[2,35],72:[2,35],75:[2,35],80:[2,35],81:[2,35],82:[2,35],83:[2,35],84:[2,35],85:[2,35]},{23:[2,36],33:[2,36],54:[2,36],65:[2,36],68:[2,36],72:[2,36],75:[2,36],80:[2,36],81:[2,36],82:[2,36],83:[2,36],84:[2,36],85:[2,36]},{23:[2,37],33:[2,37],54:[2,37],65:[2,37],68:[2,37],72:[2,37],75:[2,37],80:[2,37],81:[2,37],82:[2,37],83:[2,37],84:[2,37],85:[2,37]},{23:[2,38],33:[2,38],54:[2,38],65:[2,38],68:[2,38],72:[2,38],75:[2,38],80:[2,38],81:[2,38],82:[2,38],83:[2,38],84:[2,38],85:[2,38]},{23:[2,39],33:[2,39],54:[2,39],65:[2,39],68:[2,39],72:[2,39],75:[2,39],80:[2,39],81:[2,39],82:[2,39],83:[2,39],84:[2,39],85:[2,39]},{23:[2,43],33:[2,43],54:[2,43],65:[2,43],68:[2,43],72:[2,43],75:[2,43],80:[2,43],81:[2,43],82:[2,43],83:[2,43],84:[2,43],85:[2,43],87:[1,51]},{72:[1,35],86:52},{23:[2,45],33:[2,45],54:[2,45],65:[2,45],68:[2,45],72:[2,45],75:[2,45],80:[2,45],81:[2,45],82:[2,45],83:[2,45],84:[2,45],85:[2,45],87:[2,45]},{52:53,54:[2,82],65:[2,82],72:[2,82],80:[2,82],81:[2,82],82:[2,82],83:[2,82],84:[2,82],85:[2,82]},{25:54,38:56,39:[1,58],43:57,44:[1,59],45:55,47:[2,54]},{28:60,43:61,44:[1,59],47:[2,56]},{13:63,15:[1,20],18:[1,62]},{15:[2,48],18:[2,48]},{33:[2,86],57:64,65:[2,86],72:[2,86],80:[2,86],81:[2,86],82:[2,86],83:[2,86],84:[2,86],85:[2,86]},{33:[2,40],65:[2,40],72:[2,40],80:[2,40],81:[2,40],82:[2,40],83:[2,40],84:[2,40],85:[2,40]},{33:[2,41],65:[2,41],72:[2,41],80:[2,41],81:[2,41],82:[2,41],83:[2,41],84:[2,41],85:[2,41]},{20:65,72:[1,35],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{26:66,47:[1,67]},{30:68,33:[2,58],65:[2,58],72:[2,58],75:[2,58],80:[2,58],81:[2,58],82:[2,58],83:[2,58],84:[2,58],85:[2,58]},{33:[2,64],35:69,65:[2,64],72:[2,64],75:[2,64],80:[2,64],81:[2,64],82:[2,64],83:[2,64],84:[2,64],85:[2,64]},{21:70,23:[2,50],65:[2,50],72:[2,50],80:[2,50],81:[2,50],82:[2,50],83:[2,50],84:[2,50],85:[2,50]},{33:[2,90],61:71,65:[2,90],72:[2,90],80:[2,90],81:[2,90],82:[2,90],83:[2,90],84:[2,90],85:[2,90]},{20:75,33:[2,80],50:72,63:73,64:76,65:[1,44],69:74,70:77,71:78,72:[1,79],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{72:[1,80]},{23:[2,42],33:[2,42],54:[2,42],65:[2,42],68:[2,42],72:[2,42],75:[2,42],80:[2,42],81:[2,42],82:[2,42],83:[2,42],84:[2,42],85:[2,42],87:[1,51]},{20:75,53:81,54:[2,84],63:82,64:76,65:[1,44],69:83,70:77,71:78,72:[1,79],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{26:84,47:[1,67]},{47:[2,55]},{4:85,6:3,14:[2,46],15:[2,46],19:[2,46],29:[2,46],34:[2,46],39:[2,46],44:[2,46],47:[2,46],48:[2,46],51:[2,46],55:[2,46],60:[2,46]},{47:[2,20]},{20:86,72:[1,35],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{4:87,6:3,14:[2,46],15:[2,46],19:[2,46],29:[2,46],34:[2,46],47:[2,46],48:[2,46],51:[2,46],55:[2,46],60:[2,46]},{26:88,47:[1,67]},{47:[2,57]},{5:[2,11],14:[2,11],15:[2,11],19:[2,11],29:[2,11],34:[2,11],39:[2,11],44:[2,11],47:[2,11],48:[2,11],51:[2,11],55:[2,11],60:[2,11]},{15:[2,49],18:[2,49]},{20:75,33:[2,88],58:89,63:90,64:76,65:[1,44],69:91,70:77,71:78,72:[1,79],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{65:[2,94],66:92,68:[2,94],72:[2,94],80:[2,94],81:[2,94],82:[2,94],83:[2,94],84:[2,94],85:[2,94]},{5:[2,25],14:[2,25],15:[2,25],19:[2,25],29:[2,25],34:[2,25],39:[2,25],44:[2,25],47:[2,25],48:[2,25],51:[2,25],55:[2,25],60:[2,25]},{20:93,72:[1,35],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{20:75,31:94,33:[2,60],63:95,64:76,65:[1,44],69:96,70:77,71:78,72:[1,79],75:[2,60],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{20:75,33:[2,66],36:97,63:98,64:76,65:[1,44],69:99,70:77,71:78,72:[1,79],75:[2,66],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{20:75,22:100,23:[2,52],63:101,64:76,65:[1,44],69:102,70:77,71:78,72:[1,79],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{20:75,33:[2,92],62:103,63:104,64:76,65:[1,44],69:105,70:77,71:78,72:[1,79],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{33:[1,106]},{33:[2,79],65:[2,79],72:[2,79],80:[2,79],81:[2,79],82:[2,79],83:[2,79],84:[2,79],85:[2,79]},{33:[2,81]},{23:[2,27],33:[2,27],54:[2,27],65:[2,27],68:[2,27],72:[2,27],75:[2,27],80:[2,27],81:[2,27],82:[2,27],83:[2,27],84:[2,27],85:[2,27]},{23:[2,28],33:[2,28],54:[2,28],65:[2,28],68:[2,28],72:[2,28],75:[2,28],80:[2,28],81:[2,28],82:[2,28],83:[2,28],84:[2,28],85:[2,28]},{23:[2,30],33:[2,30],54:[2,30],68:[2,30],71:107,72:[1,108],75:[2,30]},{23:[2,98],33:[2,98],54:[2,98],68:[2,98],72:[2,98],75:[2,98]},{23:[2,45],33:[2,45],54:[2,45],65:[2,45],68:[2,45],72:[2,45],73:[1,109],75:[2,45],80:[2,45],81:[2,45],82:[2,45],83:[2,45],84:[2,45],85:[2,45],87:[2,45]},{23:[2,44],33:[2,44],54:[2,44],65:[2,44],68:[2,44],72:[2,44],75:[2,44],80:[2,44],81:[2,44],82:[2,44],83:[2,44],84:[2,44],85:[2,44],87:[2,44]},{54:[1,110]},{54:[2,83],65:[2,83],72:[2,83],80:[2,83],81:[2,83],82:[2,83],83:[2,83],84:[2,83],85:[2,83]},{54:[2,85]},{5:[2,13],14:[2,13],15:[2,13],19:[2,13],29:[2,13],34:[2,13],39:[2,13],44:[2,13],47:[2,13],48:[2,13],51:[2,13],55:[2,13],60:[2,13]},{38:56,39:[1,58],43:57,44:[1,59],45:112,46:111,47:[2,76]},{33:[2,70],40:113,65:[2,70],72:[2,70],75:[2,70],80:[2,70],81:[2,70],82:[2,70],83:[2,70],84:[2,70],85:[2,70]},{47:[2,18]},{5:[2,14],14:[2,14],15:[2,14],19:[2,14],29:[2,14],34:[2,14],39:[2,14],44:[2,14],47:[2,14],48:[2,14],51:[2,14],55:[2,14],60:[2,14]},{33:[1,114]},{33:[2,87],65:[2,87],72:[2,87],80:[2,87],81:[2,87],82:[2,87],83:[2,87],84:[2,87],85:[2,87]},{33:[2,89]},{20:75,63:116,64:76,65:[1,44],67:115,68:[2,96],69:117,70:77,71:78,72:[1,79],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{33:[1,118]},{32:119,33:[2,62],74:120,75:[1,121]},{33:[2,59],65:[2,59],72:[2,59],75:[2,59],80:[2,59],81:[2,59],82:[2,59],83:[2,59],84:[2,59],85:[2,59]},{33:[2,61],75:[2,61]},{33:[2,68],37:122,74:123,75:[1,121]},{33:[2,65],65:[2,65],72:[2,65],75:[2,65],80:[2,65],81:[2,65],82:[2,65],83:[2,65],84:[2,65],85:[2,65]},{33:[2,67],75:[2,67]},{23:[1,124]},{23:[2,51],65:[2,51],72:[2,51],80:[2,51],81:[2,51],82:[2,51],83:[2,51],84:[2,51],85:[2,51]},{23:[2,53]},{33:[1,125]},{33:[2,91],65:[2,91],72:[2,91],80:[2,91],81:[2,91],82:[2,91],83:[2,91],84:[2,91],85:[2,91]},{33:[2,93]},{5:[2,22],14:[2,22],15:[2,22],19:[2,22],29:[2,22],34:[2,22],39:[2,22],44:[2,22],47:[2,22],48:[2,22],51:[2,22],55:[2,22],60:[2,22]},{23:[2,99],33:[2,99],54:[2,99],68:[2,99],72:[2,99],75:[2,99]},{73:[1,109]},{20:75,63:126,64:76,65:[1,44],72:[1,35],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{5:[2,23],14:[2,23],15:[2,23],19:[2,23],29:[2,23],34:[2,23],39:[2,23],44:[2,23],47:[2,23],48:[2,23],51:[2,23],55:[2,23],60:[2,23]},{47:[2,19]},{47:[2,77]},{20:75,33:[2,72],41:127,63:128,64:76,65:[1,44],69:129,70:77,71:78,72:[1,79],75:[2,72],78:26,79:27,80:[1,28],81:[1,29],82:[1,30],83:[1,31],84:[1,32],85:[1,34],86:33},{5:[2,24],14:[2,24],15:[2,24],19:[2,24],29:[2,24],34:[2,24],39:[2,24],44:[2,24],47:[2,24],48:[2,24],51:[2,24],55:[2,24],60:[2,24]},{68:[1,130]},{65:[2,95],68:[2,95],72:[2,95],80:[2,95],81:[2,95],82:[2,95],83:[2,95],84:[2,95],85:[2,95]},{68:[2,97]},{5:[2,21],14:[2,21],15:[2,21],19:[2,21],29:[2,21],34:[2,21],39:[2,21],44:[2,21],47:[2,21],48:[2,21],51:[2,21],55:[2,21],60:[2,21]},{33:[1,131]},{33:[2,63]},{72:[1,133],76:132},{33:[1,134]},{33:[2,69]},{15:[2,12]},{14:[2,26],15:[2,26],19:[2,26],29:[2,26],34:[2,26],47:[2,26],48:[2,26],51:[2,26],55:[2,26],60:[2,26]},{23:[2,31],33:[2,31],54:[2,31],68:[2,31],72:[2,31],75:[2,31]},{33:[2,74],42:135,74:136,75:[1,121]},{33:[2,71],65:[2,71],72:[2,71],75:[2,71],80:[2,71],81:[2,71],82:[2,71],83:[2,71],84:[2,71],85:[2,71]},{33:[2,73],75:[2,73]},{23:[2,29],33:[2,29],54:[2,29],65:[2,29],68:[2,29],72:[2,29],75:[2,29],80:[2,29],81:[2,29],82:[2,29],83:[2,29],84:[2,29],85:[2,29]},{14:[2,15],15:[2,15],19:[2,15],29:[2,15],34:[2,15],39:[2,15],44:[2,15],47:[2,15],48:[2,15],51:[2,15],55:[2,15],60:[2,15]},{72:[1,138],77:[1,137]},{72:[2,100],77:[2,100]},{14:[2,16],15:[2,16],19:[2,16],29:[2,16],34:[2,16],44:[2,16],47:[2,16],48:[2,16],51:[2,16],55:[2,16],60:[2,16]},{33:[1,139]},{33:[2,75]},{33:[2,32]},{72:[2,101],77:[2,101]},{14:[2,17],15:[2,17],19:[2,17],29:[2,17],34:[2,17],39:[2,17],44:[2,17],47:[2,17],48:[2,17],51:[2,17],55:[2,17],60:[2,17]}],defaultActions:{4:[2,1],55:[2,55],57:[2,20],61:[2,57],74:[2,81],83:[2,85],87:[2,18],91:[2,89],102:[2,53],105:[2,93],111:[2,19],112:[2,77],117:[2,97],120:[2,63],123:[2,69],124:[2,12],136:[2,75],137:[2,32]},parseError:function(t,e){throw new Error(t)},parse:function(t){var e=this,r=[0],n=[null],i=[],a=this.table,s="",o=0,l=0,c=0;this.lexer.setInput(t),this.lexer.yy=this.yy,this.yy.lexer=this.lexer,this.yy.parser=this,
-void 0===this.lexer.yylloc&&(this.lexer.yylloc={});var p=this.lexer.yylloc;i.push(p);var h=this.lexer.options&&this.lexer.options.ranges;"function"==typeof this.yy.parseError&&(this.parseError=this.yy.parseError);for(var u,d,f,m,g,v,y,b,k,_={};;){if(f=r[r.length-1],this.defaultActions[f]?m=this.defaultActions[f]:(null!==u&&void 0!==u||(u=function(){var t;return t=e.lexer.lex()||1,"number"!=typeof t&&(t=e.symbols_[t]||t),t}()),m=a[f]&&a[f][u]),void 0===m||!m.length||!m[0]){var S="";if(!c){k=[];for(v in a[f])this.terminals_[v]&&v>2&&k.push("'"+this.terminals_[v]+"'");S=this.lexer.showPosition?"Parse error on line "+(o+1)+":\n"+this.lexer.showPosition()+"\nExpecting "+k.join(", ")+", got '"+(this.terminals_[u]||u)+"'":"Parse error on line "+(o+1)+": Unexpected "+(1==u?"end of input":"'"+(this.terminals_[u]||u)+"'"),this.parseError(S,{text:this.lexer.match,token:this.terminals_[u]||u,line:this.lexer.yylineno,loc:p,expected:k})}}if(m[0]instanceof Array&&m.length>1)throw new Error("Parse Error: multiple actions possible at state: "+f+", token: "+u);switch(m[0]){case 1:r.push(u),n.push(this.lexer.yytext),i.push(this.lexer.yylloc),r.push(m[1]),u=null,d?(u=d,d=null):(l=this.lexer.yyleng,s=this.lexer.yytext,o=this.lexer.yylineno,p=this.lexer.yylloc,c>0&&c--);break;case 2:if(y=this.productions_[m[1]][1],_.$=n[n.length-y],_._$={first_line:i[i.length-(y||1)].first_line,last_line:i[i.length-1].last_line,first_column:i[i.length-(y||1)].first_column,last_column:i[i.length-1].last_column},h&&(_._$.range=[i[i.length-(y||1)].range[0],i[i.length-1].range[1]]),void 0!==(g=this.performAction.call(_,s,l,o,this.yy,m[1],n,i)))return g;y&&(r=r.slice(0,-1*y*2),n=n.slice(0,-1*y),i=i.slice(0,-1*y)),r.push(this.productions_[m[1]][0]),n.push(_.$),i.push(_._$),b=a[r[r.length-2]][r[r.length-1]],r.push(b);break;case 3:return!0}}return!0}},r=function(){var t={EOF:1,parseError:function(t,e){if(!this.yy.parser)throw new Error(t);this.yy.parser.parseError(t,e)},setInput:function(t){return this._input=t,this._more=this._less=this.done=!1,this.yylineno=this.yyleng=0,this.yytext=this.matched=this.match="",this.conditionStack=["INITIAL"],this.yylloc={first_line:1,first_column:0,last_line:1,last_column:0},this.options.ranges&&(this.yylloc.range=[0,0]),this.offset=0,this},input:function(){var t=this._input[0];return this.yytext+=t,this.yyleng++,this.offset++,this.match+=t,this.matched+=t,t.match(/(?:\r\n?|\n).*/g)?(this.yylineno++,this.yylloc.last_line++):this.yylloc.last_column++,this.options.ranges&&this.yylloc.range[1]++,this._input=this._input.slice(1),t},unput:function(t){var e=t.length,r=t.split(/(?:\r\n?|\n)/g);this._input=t+this._input,this.yytext=this.yytext.substr(0,this.yytext.length-e-1),this.offset-=e;var n=this.match.split(/(?:\r\n?|\n)/g);this.match=this.match.substr(0,this.match.length-1),this.matched=this.matched.substr(0,this.matched.length-1),r.length-1&&(this.yylineno-=r.length-1);var i=this.yylloc.range;return this.yylloc={first_line:this.yylloc.first_line,last_line:this.yylineno+1,first_column:this.yylloc.first_column,last_column:r?(r.length===n.length?this.yylloc.first_column:0)+n[n.length-r.length].length-r[0].length:this.yylloc.first_column-e},this.options.ranges&&(this.yylloc.range=[i[0],i[0]+this.yyleng-e]),this},more:function(){return this._more=!0,this},less:function(t){this.unput(this.match.slice(t))},pastInput:function(){var t=this.matched.substr(0,this.matched.length-this.match.length);return(t.length>20?"...":"")+t.substr(-20).replace(/\n/g,"")},upcomingInput:function(){var t=this.match;return t.length<20&&(t+=this._input.substr(0,20-t.length)),(t.substr(0,20)+(t.length>20?"...":"")).replace(/\n/g,"")},showPosition:function(){var t=this.pastInput(),e=new Array(t.length+1).join("-");return t+this.upcomingInput()+"\n"+e+"^"},next:function(){if(this.done)return this.EOF;this._input||(this.done=!0);var t,e,r,n,i;this._more||(this.yytext="",this.match="");for(var a=this._currentRules(),s=0;s<a.length&&(!(r=this._input.match(this.rules[a[s]]))||e&&!(r[0].length>e[0].length)||(e=r,n=s,this.options.flex));s++);return e?(i=e[0].match(/(?:\r\n?|\n).*/g),i&&(this.yylineno+=i.length),this.yylloc={first_line:this.yylloc.last_line,last_line:this.yylineno+1,first_column:this.yylloc.last_column,last_column:i?i[i.length-1].length-i[i.length-1].match(/\r?\n?/)[0].length:this.yylloc.last_column+e[0].length},this.yytext+=e[0],this.match+=e[0],this.matches=e,this.yyleng=this.yytext.length,this.options.ranges&&(this.yylloc.range=[this.offset,this.offset+=this.yyleng]),this._more=!1,this._input=this._input.slice(e[0].length),this.matched+=e[0],t=this.performAction.call(this,this.yy,this,a[n],this.conditionStack[this.conditionStack.length-1]),this.done&&this._input&&(this.done=!1),t||void 0):""===this._input?this.EOF:this.parseError("Lexical error on line "+(this.yylineno+1)+". Unrecognized text.\n"+this.showPosition(),{text:"",token:null,line:this.yylineno})},lex:function(){var t=this.next();return void 0!==t?t:this.lex()},begin:function(t){this.conditionStack.push(t)},popState:function(){return this.conditionStack.pop()},_currentRules:function(){return this.conditions[this.conditionStack[this.conditionStack.length-1]].rules},topState:function(){return this.conditionStack[this.conditionStack.length-2]},pushState:function(t){this.begin(t)}};return t.options={},t.performAction=function(t,e,r,n){function i(t,r){return e.yytext=e.yytext.substr(t,e.yyleng-r)}switch(r){case 0:if("\\\\"===e.yytext.slice(-2)?(i(0,1),this.begin("mu")):"\\"===e.yytext.slice(-1)?(i(0,1),this.begin("emu")):this.begin("mu"),e.yytext)return 15;break;case 1:return 15;case 2:return this.popState(),15;case 3:return this.begin("raw"),15;case 4:return this.popState(),"raw"===this.conditionStack[this.conditionStack.length-1]?15:(e.yytext=e.yytext.substr(5,e.yyleng-9),"END_RAW_BLOCK");case 5:return 15;case 6:return this.popState(),14;case 7:return 65;case 8:return 68;case 9:return 19;case 10:return this.popState(),this.begin("raw"),23;case 11:return 55;case 12:return 60;case 13:return 29;case 14:return 47;case 15:case 16:return this.popState(),44;case 17:return 34;case 18:return 39;case 19:return 51;case 20:return 48;case 21:this.unput(e.yytext),this.popState(),this.begin("com");break;case 22:return this.popState(),14;case 23:return 48;case 24:return 73;case 25:case 26:return 72;case 27:return 87;case 28:break;case 29:return this.popState(),54;case 30:return this.popState(),33;case 31:return e.yytext=i(1,2).replace(/\\"/g,'"'),80;case 32:return e.yytext=i(1,2).replace(/\\'/g,"'"),80;case 33:return 85;case 34:case 35:return 82;case 36:return 83;case 37:return 84;case 38:return 81;case 39:return 75;case 40:return 77;case 41:return 72;case 42:return e.yytext=e.yytext.replace(/\\([\\\]])/g,"$1"),72;case 43:return"INVALID";case 44:return 5}},t.rules=[/^(?:[^\x00]*?(?=(\{\{)))/,/^(?:[^\x00]+)/,/^(?:[^\x00]{2,}?(?=(\{\{|\\\{\{|\\\\\{\{|$)))/,/^(?:\{\{\{\{(?=[^/]))/,/^(?:\{\{\{\{\/[^\s!"#%-,\.\/;->@\[-\^`\{-~]+(?=[=}\s\/.])\}\}\}\})/,/^(?:[^\x00]*?(?=(\{\{\{\{)))/,/^(?:[\s\S]*?--(~)?\}\})/,/^(?:\()/,/^(?:\))/,/^(?:\{\{\{\{)/,/^(?:\}\}\}\})/,/^(?:\{\{(~)?>)/,/^(?:\{\{(~)?#>)/,/^(?:\{\{(~)?#\*?)/,/^(?:\{\{(~)?\/)/,/^(?:\{\{(~)?\^\s*(~)?\}\})/,/^(?:\{\{(~)?\s*else\s*(~)?\}\})/,/^(?:\{\{(~)?\^)/,/^(?:\{\{(~)?\s*else\b)/,/^(?:\{\{(~)?\{)/,/^(?:\{\{(~)?&)/,/^(?:\{\{(~)?!--)/,/^(?:\{\{(~)?![\s\S]*?\}\})/,/^(?:\{\{(~)?\*?)/,/^(?:=)/,/^(?:\.\.)/,/^(?:\.(?=([=~}\s\/.)|])))/,/^(?:[\/.])/,/^(?:\s+)/,/^(?:\}(~)?\}\})/,/^(?:(~)?\}\})/,/^(?:"(\\["]|[^"])*")/,/^(?:'(\\[']|[^'])*')/,/^(?:@)/,/^(?:true(?=([~}\s)])))/,/^(?:false(?=([~}\s)])))/,/^(?:undefined(?=([~}\s)])))/,/^(?:null(?=([~}\s)])))/,/^(?:-?[0-9]+(?:\.[0-9]+)?(?=([~}\s)])))/,/^(?:as\s+\|)/,/^(?:\|)/,/^(?:([^\s!"#%-,\.\/;->@\[-\^`\{-~]+(?=([=~}\s\/.)|]))))/,/^(?:\[(\\\]|[^\]])*\])/,/^(?:.)/,/^(?:$)/],t.conditions={mu:{rules:[7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44],inclusive:!1},emu:{rules:[2],inclusive:!1},com:{rules:[6],inclusive:!1},raw:{rules:[3,4,5],inclusive:!1},INITIAL:{rules:[0,1,44],inclusive:!0}},t}();return e.lexer=r,t.prototype=e,e.Parser=t,new t}();e.__esModule=!0,e.default=r},function(t,e,r){"use strict";function n(){var t=arguments.length<=0||void 0===arguments[0]?{}:arguments[0];this.options=t}function i(t,e,r){void 0===e&&(e=t.length);var n=t[e-1],i=t[e-2];return n?"ContentStatement"===n.type?(i||!r?/\r?\n\s*?$/:/(^|\r?\n)\s*?$/).test(n.original):void 0:r}function a(t,e,r){void 0===e&&(e=-1);var n=t[e+1],i=t[e+2];return n?"ContentStatement"===n.type?(i||!r?/^\s*?\r?\n/:/^\s*?(\r?\n|$)/).test(n.original):void 0:r}function s(t,e,r){var n=t[null==e?0:e+1];if(n&&"ContentStatement"===n.type&&(r||!n.rightStripped)){var i=n.value;n.value=n.value.replace(r?/^\s+/:/^[ \t]*\r?\n?/,""),n.rightStripped=n.value!==i}}function o(t,e,r){var n=t[null==e?t.length-1:e-1];if(n&&"ContentStatement"===n.type&&(r||!n.leftStripped)){var i=n.value;return n.value=n.value.replace(r?/\s+$/:/[ \t]+$/,""),n.leftStripped=n.value!==i,n.leftStripped}}var l=r(1).default;e.__esModule=!0;var c=r(25),p=l(c);n.prototype=new p.default,n.prototype.Program=function(t){var e=!this.options.ignoreStandalone,r=!this.isRootSeen;this.isRootSeen=!0;for(var n=t.body,l=0,c=n.length;l<c;l++){var p=n[l],h=this.accept(p);if(h){var u=i(n,l,r),d=a(n,l,r),f=h.openStandalone&&u,m=h.closeStandalone&&d,g=h.inlineStandalone&&u&&d;h.close&&s(n,l,!0),h.open&&o(n,l,!0),e&&g&&(s(n,l),o(n,l)&&"PartialStatement"===p.type&&(p.indent=/([ \t]+$)/.exec(n[l-1].original)[1])),e&&f&&(s((p.program||p.inverse).body),o(n,l)),e&&m&&(s(n,l),o((p.inverse||p.program).body))}}return t},n.prototype.BlockStatement=n.prototype.DecoratorBlock=n.prototype.PartialBlockStatement=function(t){this.accept(t.program),this.accept(t.inverse);var e=t.program||t.inverse,r=t.program&&t.inverse,n=r,l=r;if(r&&r.chained)for(n=r.body[0].program;l.chained;)l=l.body[l.body.length-1].program;var c={open:t.openStrip.open,close:t.closeStrip.close,openStandalone:a(e.body),closeStandalone:i((n||e).body)};if(t.openStrip.close&&s(e.body,null,!0),r){var p=t.inverseStrip;p.open&&o(e.body,null,!0),p.close&&s(n.body,null,!0),t.closeStrip.open&&o(l.body,null,!0),!this.options.ignoreStandalone&&i(e.body)&&a(n.body)&&(o(e.body),s(n.body))}else t.closeStrip.open&&o(e.body,null,!0);return c},n.prototype.Decorator=n.prototype.MustacheStatement=function(t){return t.strip},n.prototype.PartialStatement=n.prototype.CommentStatement=function(t){var e=t.strip||{};return{inlineStandalone:!0,open:e.open,close:e.close}},e.default=n,t.exports=e.default},function(t,e,r){"use strict";function n(){this.parents=[]}function i(t){this.acceptRequired(t,"path"),this.acceptArray(t.params),this.acceptKey(t,"hash")}function a(t){i.call(this,t),this.acceptKey(t,"program"),this.acceptKey(t,"inverse")}function s(t){this.acceptRequired(t,"name"),this.acceptArray(t.params),this.acceptKey(t,"hash")}var o=r(1).default;e.__esModule=!0;var l=r(6),c=o(l);n.prototype={constructor:n,mutating:!1,acceptKey:function(t,e){var r=this.accept(t[e]);if(this.mutating){if(r&&!n.prototype[r.type])throw new c.default('Unexpected node type "'+r.type+'" found when accepting '+e+" on "+t.type);t[e]=r}},acceptRequired:function(t,e){if(this.acceptKey(t,e),!t[e])throw new c.default(t.type+" requires "+e)},acceptArray:function(t){for(var e=0,r=t.length;e<r;e++)this.acceptKey(t,e),t[e]||(t.splice(e,1),e--,r--)},accept:function(t){if(t){if(!this[t.type])throw new c.default("Unknown type: "+t.type,t);this.current&&this.parents.unshift(this.current),this.current=t;var e=this[t.type](t);return this.current=this.parents.shift(),!this.mutating||e?e:!1!==e?t:void 0}},Program:function(t){this.acceptArray(t.body)},MustacheStatement:i,Decorator:i,BlockStatement:a,DecoratorBlock:a,PartialStatement:s,PartialBlockStatement:function(t){s.call(this,t),this.acceptKey(t,"program")},ContentStatement:function(){},CommentStatement:function(){},SubExpression:i,PathExpression:function(){},StringLiteral:function(){},NumberLiteral:function(){},BooleanLiteral:function(){},UndefinedLiteral:function(){},NullLiteral:function(){},Hash:function(t){this.acceptArray(t.pairs)},HashPair:function(t){this.acceptRequired(t,"value")}},e.default=n,t.exports=e.default},function(t,e,r){"use strict";function n(t,e){if(e=e.path?e.path.original:e,t.path.original!==e){var r={loc:t.path.loc};throw new g.default(t.path.original+" doesn't match "+e,r)}}function i(t,e){this.source=t,this.start={line:e.first_line,column:e.first_column},this.end={line:e.last_line,column:e.last_column}}function a(t){return/^\[.*\]$/.test(t)?t.substr(1,t.length-2):t}function s(t,e){return{open:"~"===t.charAt(2),close:"~"===e.charAt(e.length-3)}}function o(t){return t.replace(/^\{\{~?\!-?-?/,"").replace(/-?-?~?\}\}$/,"")}function l(t,e,r){r=this.locInfo(r);for(var n=t?"@":"",i=[],a=0,s="",o=0,l=e.length;o<l;o++){var c=e[o].part,p=e[o].original!==c;if(n+=(e[o].separator||"")+c,p||".."!==c&&"."!==c&&"this"!==c)i.push(c);else{if(i.length>0)throw new g.default("Invalid path: "+n,{loc:r});".."===c&&(a++,s+="../")}}return{type:"PathExpression",data:t,depth:a,parts:i,original:n,loc:r}}function c(t,e,r,n,i,a){var s=n.charAt(3)||n.charAt(2),o="{"!==s&&"&"!==s;return{type:/\*/.test(n)?"Decorator":"MustacheStatement",path:t,params:e,hash:r,escaped:o,strip:i,loc:this.locInfo(a)}}function p(t,e,r,i){n(t,r),i=this.locInfo(i);var a={type:"Program",body:e,strip:{},loc:i};return{type:"BlockStatement",path:t.path,params:t.params,hash:t.hash,program:a,openStrip:{},inverseStrip:{},closeStrip:{},loc:i}}function h(t,e,r,i,a,s){i&&i.path&&n(t,i);var o=/\*/.test(t.open);e.blockParams=t.blockParams;var l=void 0,c=void 0;if(r){if(o)throw new g.default("Unexpected inverse block on decorator",r);r.chain&&(r.program.body[0].closeStrip=i.strip),c=r.strip,l=r.program}return a&&(a=l,l=e,e=a),{type:o?"DecoratorBlock":"BlockStatement",path:t.path,params:t.params,hash:t.hash,program:e,inverse:l,openStrip:t.strip,inverseStrip:c,closeStrip:i&&i.strip,loc:this.locInfo(s)}}function u(t,e){if(!e&&t.length){var r=t[0].loc,n=t[t.length-1].loc;r&&n&&(e={source:r.source,start:{line:r.start.line,column:r.start.column},end:{line:n.end.line,column:n.end.column}})}return{type:"Program",body:t,strip:{},loc:e}}function d(t,e,r,i){return n(t,r),{type:"PartialBlockStatement",name:t.path,params:t.params,hash:t.hash,program:e,openStrip:t.strip,closeStrip:r&&r.strip,loc:this.locInfo(i)}}var f=r(1).default;e.__esModule=!0,e.SourceLocation=i,e.id=a,e.stripFlags=s,e.stripComment=o,e.preparePath=l,e.prepareMustache=c,e.prepareRawBlock=p,e.prepareBlock=h,e.prepareProgram=u,e.preparePartialBlock=d;var m=r(6),g=f(m)},function(t,e,r){"use strict";function n(){}function i(t,e,r){if(null==t||"string"!=typeof t&&"Program"!==t.type)throw new p.default("You must pass a string or Handlebars AST to Handlebars.precompile. You passed "+t);e=e||{},"data"in e||(e.data=!0),e.compat&&(e.useDepths=!0);var n=r.parse(t,e),i=(new r.Compiler).compile(n,e);return(new r.JavaScriptCompiler).compile(i,e)}function a(t,e,r){function n(){var n=r.parse(t,e),i=(new r.Compiler).compile(n,e),a=(new r.JavaScriptCompiler).compile(i,e,void 0,!0);return r.template(a)}function i(t,e){return a||(a=n()),a.call(this,t,e)}if(void 0===e&&(e={}),null==t||"string"!=typeof t&&"Program"!==t.type)throw new p.default("You must pass a string or Handlebars AST to Handlebars.compile. You passed "+t);"data"in e||(e.data=!0),e.compat&&(e.useDepths=!0);var a=void 0;return i._setup=function(t){return a||(a=n()),a._setup(t)},i._child=function(t,e,r,i){return a||(a=n()),a._child(t,e,r,i)},i}function s(t,e){if(t===e)return!0;if(h.isArray(t)&&h.isArray(e)&&t.length===e.length){for(var r=0;r<t.length;r++)if(!s(t[r],e[r]))return!1;return!0}}function o(t){if(!t.path.parts){var e=t.path;t.path={type:"PathExpression",data:!1,depth:0,parts:[e.original+""],original:e.original+"",loc:e.loc}}}var l=r(1).default;e.__esModule=!0,e.Compiler=n,e.precompile=i,e.compile=a;var c=r(6),p=l(c),h=r(5),u=r(21),d=l(u),f=[].slice;n.prototype={compiler:n,equals:function(t){var e=this.opcodes.length;if(t.opcodes.length!==e)return!1;for(var r=0;r<e;r++){var n=this.opcodes[r],i=t.opcodes[r];if(n.opcode!==i.opcode||!s(n.args,i.args))return!1}e=this.children.length;for(var r=0;r<e;r++)if(!this.children[r].equals(t.children[r]))return!1;return!0},guid:0,compile:function(t,e){this.sourceNode=[],this.opcodes=[],this.children=[],this.options=e,this.stringParams=e.stringParams,this.trackIds=e.trackIds,e.blockParams=e.blockParams||[];var r=e.knownHelpers;if(e.knownHelpers={helperMissing:!0,blockHelperMissing:!0,each:!0,if:!0,unless:!0,with:!0,log:!0,lookup:!0},r)for(var n in r)n in r&&(e.knownHelpers[n]=r[n]);return this.accept(t)},compileProgram:function(t){var e=new this.compiler,r=e.compile(t,this.options),n=this.guid++;return this.usePartial=this.usePartial||r.usePartial,this.children[n]=r,this.useDepths=this.useDepths||r.useDepths,n},accept:function(t){if(!this[t.type])throw new p.default("Unknown type: "+t.type,t);this.sourceNode.unshift(t);var e=this[t.type](t);return this.sourceNode.shift(),e},Program:function(t){this.options.blockParams.unshift(t.blockParams);for(var e=t.body,r=e.length,n=0;n<r;n++)this.accept(e[n]);return this.options.blockParams.shift(),this.isSimple=1===r,this.blockParams=t.blockParams?t.blockParams.length:0,this},BlockStatement:function(t){o(t);var e=t.program,r=t.inverse;e=e&&this.compileProgram(e),r=r&&this.compileProgram(r);var n=this.classifySexpr(t);"helper"===n?this.helperSexpr(t,e,r):"simple"===n?(this.simpleSexpr(t),this.opcode("pushProgram",e),this.opcode("pushProgram",r),this.opcode("emptyHash"),this.opcode("blockValue",t.path.original)):(this.ambiguousSexpr(t,e,r),this.opcode("pushProgram",e),this.opcode("pushProgram",r),this.opcode("emptyHash"),this.opcode("ambiguousBlockValue")),this.opcode("append")},DecoratorBlock:function(t){var e=t.program&&this.compileProgram(t.program),r=this.setupFullMustacheParams(t,e,void 0),n=t.path;this.useDecorators=!0,this.opcode("registerDecorator",r.length,n.original)},PartialStatement:function(t){this.usePartial=!0;var e=t.program;e&&(e=this.compileProgram(t.program));var r=t.params;if(r.length>1)throw new p.default("Unsupported number of partial arguments: "+r.length,t);r.length||(this.options.explicitPartialContext?this.opcode("pushLiteral","undefined"):r.push({type:"PathExpression",parts:[],depth:0}));var n=t.name.original,i="SubExpression"===t.name.type;i&&this.accept(t.name),this.setupFullMustacheParams(t,e,void 0,!0);var a=t.indent||"";this.options.preventIndent&&a&&(this.opcode("appendContent",a),a=""),this.opcode("invokePartial",i,n,a),this.opcode("append")},PartialBlockStatement:function(t){this.PartialStatement(t)},MustacheStatement:function(t){this.SubExpression(t),t.escaped&&!this.options.noEscape?this.opcode("appendEscaped"):this.opcode("append")},Decorator:function(t){this.DecoratorBlock(t)},ContentStatement:function(t){t.value&&this.opcode("appendContent",t.value)},CommentStatement:function(){},SubExpression:function(t){o(t);var e=this.classifySexpr(t);"simple"===e?this.simpleSexpr(t):"helper"===e?this.helperSexpr(t):this.ambiguousSexpr(t)},ambiguousSexpr:function(t,e,r){var n=t.path,i=n.parts[0],a=null!=e||null!=r;this.opcode("getContext",n.depth),this.opcode("pushProgram",e),this.opcode("pushProgram",r),n.strict=!0,this.accept(n),this.opcode("invokeAmbiguous",i,a)},simpleSexpr:function(t){var e=t.path;e.strict=!0,this.accept(e),this.opcode("resolvePossibleLambda")},helperSexpr:function(t,e,r){var n=this.setupFullMustacheParams(t,e,r),i=t.path,a=i.parts[0];if(this.options.knownHelpers[a])this.opcode("invokeKnownHelper",n.length,a);else{if(this.options.knownHelpersOnly)throw new p.default("You specified knownHelpersOnly, but used the unknown helper "+a,t);i.strict=!0,i.falsy=!0,this.accept(i),this.opcode("invokeHelper",n.length,i.original,d.default.helpers.simpleId(i))}},PathExpression:function(t){this.addDepth(t.depth),this.opcode("getContext",t.depth);var e=t.parts[0],r=d.default.helpers.scopedId(t),n=!t.depth&&!r&&this.blockParamIndex(e);n?this.opcode("lookupBlockParam",n,t.parts):e?t.data?(this.options.data=!0,this.opcode("lookupData",t.depth,t.parts,t.strict)):this.opcode("lookupOnContext",t.parts,t.falsy,t.strict,r):this.opcode("pushContext")},StringLiteral:function(t){this.opcode("pushString",t.value)},NumberLiteral:function(t){this.opcode("pushLiteral",t.value)},BooleanLiteral:function(t){this.opcode("pushLiteral",t.value)},UndefinedLiteral:function(){this.opcode("pushLiteral","undefined")},NullLiteral:function(){this.opcode("pushLiteral","null")},Hash:function(t){var e=t.pairs,r=0,n=e.length;for(this.opcode("pushHash");r<n;r++)this.pushParam(e[r].value);for(;r--;)this.opcode("assignToHash",e[r].key);this.opcode("popHash")},opcode:function(t){this.opcodes.push({opcode:t,args:f.call(arguments,1),loc:this.sourceNode[0].loc})},addDepth:function(t){t&&(this.useDepths=!0)},classifySexpr:function(t){var e=d.default.helpers.simpleId(t.path),r=e&&!!this.blockParamIndex(t.path.parts[0]),n=!r&&d.default.helpers.helperExpression(t),i=!r&&(n||e);if(i&&!n){var a=t.path.parts[0],s=this.options;s.knownHelpers[a]?n=!0:s.knownHelpersOnly&&(i=!1)}return n?"helper":i?"ambiguous":"simple"},pushParams:function(t){for(var e=0,r=t.length;e<r;e++)this.pushParam(t[e])},pushParam:function(t){var e=null!=t.value?t.value:t.original||"";if(this.stringParams)e.replace&&(e=e.replace(/^(\.?\.\/)*/g,"").replace(/\//g,".")),t.depth&&this.addDepth(t.depth),this.opcode("getContext",t.depth||0),this.opcode("pushStringParam",e,t.type),"SubExpression"===t.type&&this.accept(t);else{if(this.trackIds){var r=void 0;if(!t.parts||d.default.helpers.scopedId(t)||t.depth||(r=this.blockParamIndex(t.parts[0])),r){var n=t.parts.slice(1).join(".");this.opcode("pushId","BlockParam",r,n)}else e=t.original||e,e.replace&&(e=e.replace(/^this(?:\.|$)/,"").replace(/^\.\//,"").replace(/^\.$/,"")),this.opcode("pushId",t.type,e)}this.accept(t)}},setupFullMustacheParams:function(t,e,r,n){var i=t.params;return this.pushParams(i),this.opcode("pushProgram",e),this.opcode("pushProgram",r),t.hash?this.accept(t.hash):this.opcode("emptyHash",n),i},blockParamIndex:function(t){for(var e=0,r=this.options.blockParams.length;e<r;e++){var n=this.options.blockParams[e],i=n&&h.indexOf(n,t);if(n&&i>=0)return[e,i]}}}},function(t,e,r){"use strict";function n(t){this.value=t}function i(){}function a(t,e,r,n){var i=e.popStack(),a=0,s=r.length;for(t&&s--;a<s;a++)i=e.nameLookup(i,r[a],n);return t?[e.aliasable("container.strict"),"(",i,", ",e.quotedString(r[a]),")"]:i}var s=r(1).default;e.__esModule=!0;var o=r(4),l=r(6),c=s(l),p=r(5),h=r(29),u=s(h);i.prototype={nameLookup:function(t,e){return i.isValidJavaScriptVariableName(e)?[t,".",e]:[t,"[",JSON.stringify(e),"]"]},depthedLookup:function(t){return[this.aliasable("container.lookup"),'(depths, "',t,'")']},compilerInfo:function(){var t=o.COMPILER_REVISION;return[t,o.REVISION_CHANGES[t]]},appendToBuffer:function(t,e,r){return p.isArray(t)||(t=[t]),t=this.source.wrap(t,e),this.environment.isSimple?["return ",t,";"]:r?["buffer += ",t,";"]:(t.appendToBuffer=!0,t)},initializeBuffer:function(){return this.quotedString("")},compile:function(t,e,r,n){this.environment=t,this.options=e,this.stringParams=this.options.stringParams,this.trackIds=this.options.trackIds,this.precompile=!n,this.name=this.environment.name,this.isChild=!!r,this.context=r||{decorators:[],programs:[],environments:[]},this.preamble(),this.stackSlot=0,this.stackVars=[],this.aliases={},this.registers={list:[]},this.hashes=[],this.compileStack=[],this.inlineStack=[],this.blockParams=[],this.compileChildren(t,e),this.useDepths=this.useDepths||t.useDepths||t.useDecorators||this.options.compat,this.useBlockParams=this.useBlockParams||t.useBlockParams;var i=t.opcodes,a=void 0,s=void 0,o=void 0,l=void 0;for(o=0,l=i.length;o<l;o++)a=i[o],this.source.currentLocation=a.loc,s=s||a.loc,this[a.opcode].apply(this,a.args);if(this.source.currentLocation=s,this.pushSource(""),this.stackSlot||this.inlineStack.length||this.compileStack.length)throw new c.default("Compile completed with content left on stack");this.decorators.isEmpty()?this.decorators=void 0:(this.useDecorators=!0,this.decorators.prepend("var decorators = container.decorators;\n"),this.decorators.push("return fn;"),n?this.decorators=Function.apply(this,["fn","props","container","depth0","data","blockParams","depths",this.decorators.merge()]):(this.decorators.prepend("function(fn, props, container, depth0, data, blockParams, depths) {\n"),this.decorators.push("}\n"),this.decorators=this.decorators.merge()));var p=this.createFunctionContext(n);if(this.isChild)return p;var h={compiler:this.compilerInfo(),main:p};this.decorators&&(h.main_d=this.decorators,h.useDecorators=!0);var u=this.context,d=u.programs,f=u.decorators;for(o=0,l=d.length;o<l;o++)d[o]&&(h[o]=d[o],f[o]&&(h[o+"_d"]=f[o],h.useDecorators=!0));return this.environment.usePartial&&(h.usePartial=!0),this.options.data&&(h.useData=!0),this.useDepths&&(h.useDepths=!0),this.useBlockParams&&(h.useBlockParams=!0),this.options.compat&&(h.compat=!0),n?h.compilerOptions=this.options:(h.compiler=JSON.stringify(h.compiler),this.source.currentLocation={start:{line:1,column:0}},h=this.objectLiteral(h),e.srcName?(h=h.toStringWithSourceMap({file:e.destName}),h.map=h.map&&h.map.toString()):h=h.toString()),h},preamble:function(){this.lastContext=0,this.source=new u.default(this.options.srcName),this.decorators=new u.default(this.options.srcName)},createFunctionContext:function(t){var e="",r=this.stackVars.concat(this.registers.list);r.length>0&&(e+=", "+r.join(", "));var n=0;for(var i in this.aliases){var a=this.aliases[i];this.aliases.hasOwnProperty(i)&&a.children&&a.referenceCount>1&&(e+=", alias"+ ++n+"="+i,a.children[0]="alias"+n)}var s=["container","depth0","helpers","partials","data"];(this.useBlockParams||this.useDepths)&&s.push("blockParams"),this.useDepths&&s.push("depths");var o=this.mergeSource(e);return t?(s.push(o),Function.apply(this,s)):this.source.wrap(["function(",s.join(","),") {\n  ",o,"}"])},mergeSource:function(t){var e=this.environment.isSimple,r=!this.forceBuffer,n=void 0,i=void 0,a=void 0,s=void 0;return this.source.each(function(t){t.appendToBuffer?(a?t.prepend("  + "):a=t,s=t):(a&&(i?a.prepend("buffer += "):n=!0,s.add(";"),a=s=void 0),i=!0,e||(r=!1))}),r?a?(a.prepend("return "),s.add(";")):i||this.source.push('return "";'):(t+=", buffer = "+(n?"":this.initializeBuffer()),a?(a.prepend("return buffer + "),s.add(";")):this.source.push("return buffer;")),t&&this.source.prepend("var "+t.substring(2)+(n?"":";\n")),this.source.merge()},blockValue:function(t){var e=this.aliasable("helpers.blockHelperMissing"),r=[this.contextName(0)];this.setupHelperArgs(t,0,r);var n=this.popStack();r.splice(1,0,n),this.push(this.source.functionCall(e,"call",r))},ambiguousBlockValue:function(){var t=this.aliasable("helpers.blockHelperMissing"),e=[this.contextName(0)];this.setupHelperArgs("",0,e,!0),this.flushInline();var r=this.topStack();e.splice(1,0,r),this.pushSource(["if (!",this.lastHelper,") { ",r," = ",this.source.functionCall(t,"call",e),"}"])},appendContent:function(t){this.pendingContent?t=this.pendingContent+t:this.pendingLocation=this.source.currentLocation,this.pendingContent=t},append:function(){if(this.isInline())this.replaceStack(function(t){return[" != null ? ",t,' : ""']}),this.pushSource(this.appendToBuffer(this.popStack()));else{var t=this.popStack();this.pushSource(["if (",t," != null) { ",this.appendToBuffer(t,void 0,!0)," }"]),this.environment.isSimple&&this.pushSource(["else { ",this.appendToBuffer("''",void 0,!0)," }"])}},appendEscaped:function(){this.pushSource(this.appendToBuffer([this.aliasable("container.escapeExpression"),"(",this.popStack(),")"]))},getContext:function(t){this.lastContext=t},pushContext:function(){this.pushStackLiteral(this.contextName(this.lastContext))},lookupOnContext:function(t,e,r,n){var i=0;n||!this.options.compat||this.lastContext?this.pushContext():this.push(this.depthedLookup(t[i++])),this.resolvePath("context",t,i,e,r)},lookupBlockParam:function(t,e){this.useBlockParams=!0,this.push(["blockParams[",t[0],"][",t[1],"]"]),this.resolvePath("context",e,1)},lookupData:function(t,e,r){t?this.pushStackLiteral("container.data(data, "+t+")"):this.pushStackLiteral("data"),this.resolvePath("data",e,0,!0,r)},resolvePath:function(t,e,r,n,i){var s=this;if(this.options.strict||this.options.assumeObjects)return void this.push(a(this.options.strict&&i,this,e,t));for(var o=e.length;r<o;r++)this.replaceStack(function(i){var a=s.nameLookup(i,e[r],t);return n?[" && ",a]:[" != null ? ",a," : ",i]})},resolvePossibleLambda:function(){this.push([this.aliasable("container.lambda"),"(",this.popStack(),", ",this.contextName(0),")"])},pushStringParam:function(t,e){this.pushContext(),this.pushString(e),"SubExpression"!==e&&("string"==typeof t?this.pushString(t):this.pushStackLiteral(t))},emptyHash:function(t){this.trackIds&&this.push("{}"),this.stringParams&&(this.push("{}"),this.push("{}")),this.pushStackLiteral(t?"undefined":"{}")},pushHash:function(){this.hash&&this.hashes.push(this.hash),this.hash={values:[],types:[],contexts:[],ids:[]}},popHash:function(){var t=this.hash;this.hash=this.hashes.pop(),this.trackIds&&this.push(this.objectLiteral(t.ids)),this.stringParams&&(this.push(this.objectLiteral(t.contexts)),this.push(this.objectLiteral(t.types))),this.push(this.objectLiteral(t.values))},pushString:function(t){this.pushStackLiteral(this.quotedString(t))},pushLiteral:function(t){this.pushStackLiteral(t)},pushProgram:function(t){null!=t?this.pushStackLiteral(this.programExpression(t)):this.pushStackLiteral(null)},registerDecorator:function(t,e){var r=this.nameLookup("decorators",e,"decorator"),n=this.setupHelperArgs(e,t);this.decorators.push(["fn = ",this.decorators.functionCall(r,"",["fn","props","container",n])," || fn;"])},invokeHelper:function(t,e,r){var n=this.popStack(),i=this.setupHelper(t,e),a=r?[i.name," || "]:"",s=["("].concat(a,n);this.options.strict||s.push(" || ",this.aliasable("helpers.helperMissing")),s.push(")"),this.push(this.source.functionCall(s,"call",i.callParams))},invokeKnownHelper:function(t,e){var r=this.setupHelper(t,e);this.push(this.source.functionCall(r.name,"call",r.callParams))},invokeAmbiguous:function(t,e){this.useRegister("helper");var r=this.popStack();this.emptyHash();var n=this.setupHelper(0,t,e),i=this.lastHelper=this.nameLookup("helpers",t,"helper"),a=["(","(helper = ",i," || ",r,")"];this.options.strict||(a[0]="(helper = ",a.push(" != null ? helper : ",this.aliasable("helpers.helperMissing"))),this.push(["(",a,n.paramsInit?["),(",n.paramsInit]:[],"),","(typeof helper === ",this.aliasable('"function"')," ? ",this.source.functionCall("helper","call",n.callParams)," : helper))"])},invokePartial:function(t,e,r){var n=[],i=this.setupParams(e,1,n);t&&(e=this.popStack(),delete i.name),r&&(i.indent=JSON.stringify(r)),i.helpers="helpers",i.partials="partials",i.decorators="container.decorators",t?n.unshift(e):n.unshift(this.nameLookup("partials",e,"partial")),this.options.compat&&(i.depths="depths"),i=this.objectLiteral(i),n.push(i),this.push(this.source.functionCall("container.invokePartial","",n))},assignToHash:function(t){var e=this.popStack(),r=void 0,n=void 0,i=void 0;this.trackIds&&(i=this.popStack()),this.stringParams&&(n=this.popStack(),r=this.popStack());var a=this.hash;r&&(a.contexts[t]=r),n&&(a.types[t]=n),i&&(a.ids[t]=i),a.values[t]=e},pushId:function(t,e,r){"BlockParam"===t?this.pushStackLiteral("blockParams["+e[0]+"].path["+e[1]+"]"+(r?" + "+JSON.stringify("."+r):"")):"PathExpression"===t?this.pushString(e):"SubExpression"===t?this.pushStackLiteral("true"):this.pushStackLiteral("null")},compiler:i,compileChildren:function(t,e){for(var r=t.children,n=void 0,i=void 0,a=0,s=r.length;a<s;a++){n=r[a],i=new this.compiler;var o=this.matchExistingProgram(n);null==o?(this.context.programs.push(""),o=this.context.programs.length,n.index=o,
-n.name="program"+o,this.context.programs[o]=i.compile(n,e,this.context,!this.precompile),this.context.decorators[o]=i.decorators,this.context.environments[o]=n,this.useDepths=this.useDepths||i.useDepths,this.useBlockParams=this.useBlockParams||i.useBlockParams):(n.index=o,n.name="program"+o,this.useDepths=this.useDepths||n.useDepths,this.useBlockParams=this.useBlockParams||n.useBlockParams)}},matchExistingProgram:function(t){for(var e=0,r=this.context.environments.length;e<r;e++){var n=this.context.environments[e];if(n&&n.equals(t))return e}},programExpression:function(t){var e=this.environment.children[t],r=[e.index,"data",e.blockParams];return(this.useBlockParams||this.useDepths)&&r.push("blockParams"),this.useDepths&&r.push("depths"),"container.program("+r.join(", ")+")"},useRegister:function(t){this.registers[t]||(this.registers[t]=!0,this.registers.list.push(t))},push:function(t){return t instanceof n||(t=this.source.wrap(t)),this.inlineStack.push(t),t},pushStackLiteral:function(t){this.push(new n(t))},pushSource:function(t){this.pendingContent&&(this.source.push(this.appendToBuffer(this.source.quotedString(this.pendingContent),this.pendingLocation)),this.pendingContent=void 0),t&&this.source.push(t)},replaceStack:function(t){var e=["("],r=void 0,i=void 0,a=void 0;if(!this.isInline())throw new c.default("replaceStack on non-inline");var s=this.popStack(!0);if(s instanceof n)r=[s.value],e=["(",r],a=!0;else{i=!0;var o=this.incrStack();e=["((",this.push(o)," = ",s,")"],r=this.topStack()}var l=t.call(this,r);a||this.popStack(),i&&this.stackSlot--,this.push(e.concat(l,")"))},incrStack:function(){return this.stackSlot++,this.stackSlot>this.stackVars.length&&this.stackVars.push("stack"+this.stackSlot),this.topStackName()},topStackName:function(){return"stack"+this.stackSlot},flushInline:function(){var t=this.inlineStack;this.inlineStack=[];for(var e=0,r=t.length;e<r;e++){var i=t[e];if(i instanceof n)this.compileStack.push(i);else{var a=this.incrStack();this.pushSource([a," = ",i,";"]),this.compileStack.push(a)}}},isInline:function(){return this.inlineStack.length},popStack:function(t){var e=this.isInline(),r=(e?this.inlineStack:this.compileStack).pop();if(!t&&r instanceof n)return r.value;if(!e){if(!this.stackSlot)throw new c.default("Invalid stack pop");this.stackSlot--}return r},topStack:function(){var t=this.isInline()?this.inlineStack:this.compileStack,e=t[t.length-1];return e instanceof n?e.value:e},contextName:function(t){return this.useDepths&&t?"depths["+t+"]":"depth"+t},quotedString:function(t){return this.source.quotedString(t)},objectLiteral:function(t){return this.source.objectLiteral(t)},aliasable:function(t){var e=this.aliases[t];return e?(e.referenceCount++,e):(e=this.aliases[t]=this.source.wrap(t),e.aliasable=!0,e.referenceCount=1,e)},setupHelper:function(t,e,r){var n=[];return{params:n,paramsInit:this.setupHelperArgs(e,t,n,r),name:this.nameLookup("helpers",e,"helper"),callParams:[this.aliasable(this.contextName(0)+" != null ? "+this.contextName(0)+" : {}")].concat(n)}},setupParams:function(t,e,r){var n={},i=[],a=[],s=[],o=!r,l=void 0;o&&(r=[]),n.name=this.quotedString(t),n.hash=this.popStack(),this.trackIds&&(n.hashIds=this.popStack()),this.stringParams&&(n.hashTypes=this.popStack(),n.hashContexts=this.popStack());var c=this.popStack(),p=this.popStack();(p||c)&&(n.fn=p||"container.noop",n.inverse=c||"container.noop");for(var h=e;h--;)l=this.popStack(),r[h]=l,this.trackIds&&(s[h]=this.popStack()),this.stringParams&&(a[h]=this.popStack(),i[h]=this.popStack());return o&&(n.args=this.source.generateArray(r)),this.trackIds&&(n.ids=this.source.generateArray(s)),this.stringParams&&(n.types=this.source.generateArray(a),n.contexts=this.source.generateArray(i)),this.options.data&&(n.data="data"),this.useBlockParams&&(n.blockParams="blockParams"),n},setupHelperArgs:function(t,e,r,n){var i=this.setupParams(t,e,r);return i=this.objectLiteral(i),n?(this.useRegister("options"),r.push("options"),["options=",i]):r?(r.push(i),""):i}},function(){for(var t="break else new var case finally return void catch for switch while continue function this with default if throw delete in try do instanceof typeof abstract enum int short boolean export interface static byte extends long super char final native synchronized class float package throws const goto private transient debugger implements protected volatile double import public let yield await null true false".split(" "),e=i.RESERVED_WORDS={},r=0,n=t.length;r<n;r++)e[t[r]]=!0}(),i.isValidJavaScriptVariableName=function(t){return!i.RESERVED_WORDS[t]&&/^[a-zA-Z_$][0-9a-zA-Z_$]*$/.test(t)},e.default=i,t.exports=e.default},function(t,e,r){"use strict";function n(t,e,r){if(a.isArray(t)){for(var n=[],i=0,s=t.length;i<s;i++)n.push(e.wrap(t[i],r));return n}return"boolean"==typeof t||"number"==typeof t?t+"":t}function i(t){this.srcFile=t,this.source=[]}e.__esModule=!0;var a=r(5),s=void 0;try{}catch(t){}s||(s=function(t,e,r,n){this.src="",n&&this.add(n)},s.prototype={add:function(t){a.isArray(t)&&(t=t.join("")),this.src+=t},prepend:function(t){a.isArray(t)&&(t=t.join("")),this.src=t+this.src},toStringWithSourceMap:function(){return{code:this.toString()}},toString:function(){return this.src}}),i.prototype={isEmpty:function(){return!this.source.length},prepend:function(t,e){this.source.unshift(this.wrap(t,e))},push:function(t,e){this.source.push(this.wrap(t,e))},merge:function(){var t=this.empty();return this.each(function(e){t.add(["  ",e,"\n"])}),t},each:function(t){for(var e=0,r=this.source.length;e<r;e++)t(this.source[e])},empty:function(){var t=this.currentLocation||{start:{}};return new s(t.start.line,t.start.column,this.srcFile)},wrap:function(t){var e=arguments.length<=1||void 0===arguments[1]?this.currentLocation||{start:{}}:arguments[1];return t instanceof s?t:(t=n(t,this,e),new s(e.start.line,e.start.column,this.srcFile,t))},functionCall:function(t,e,r){return r=this.generateList(r),this.wrap([t,e?"."+e+"(":"(",r,")"])},quotedString:function(t){return'"'+(t+"").replace(/\\/g,"\\\\").replace(/"/g,'\\"').replace(/\n/g,"\\n").replace(/\r/g,"\\r").replace(/\u2028/g,"\\u2028").replace(/\u2029/g,"\\u2029")+'"'},objectLiteral:function(t){var e=[];for(var r in t)if(t.hasOwnProperty(r)){var i=n(t[r],this);"undefined"!==i&&e.push([this.quotedString(r),":",i])}var a=this.generateList(e);return a.prepend("{"),a.add("}"),a},generateList:function(t){for(var e=this.empty(),r=0,i=t.length;r<i;r++)r&&e.add(","),e.add(n(t[r],this));return e},generateArray:function(t){var e=this.generateList(t);return e.prepend("["),e.add("]"),e}},e.default=i,t.exports=e.default}])}),define("handlebars",function(){}),function(t,e,r){e(function(){function t(t){var i=this;i.button=e(t);var a="#"+i.button.attr("id");i.isInTinyMce=Boolean("browse-internal"===i.button.attr("name")),i.isInTinyMce?(e(document).off("mouseup",a),e(document).on("mouseup",a,i.openOverlay.bind(null,i))):(e(document).off("click",a),e(document).on("click",a,i.openOverlay.bind(null,i))),e(window).one("resize",i.resize),e(document).one("keydown",function(t){27==t.which&&(t.stopPropagation(),n(t))}),i.field=i.button.closest(".referencewidgetField"),i.name=i.field.data("fieldname"),i.translations=i.button.closest(".referencewidget").data("trans"),i.request_data={},i.widget_url="",i.field_id="",i.list_template="",i.checkbox_template="",i.request_path=i.field.data("path"),i.sel_type="",i.page=1,i.term="",i.search_current_path=0,i.list_template=r.compile(e("#listing-template",i.field).html()),i.checkbox_template=r.compile(e("#checkbox-template",i.field).html()),i.sel_type=i.button.closest(".referencewidget").data("type"),i.loadSelectedItems(i),e(document).on("OverlayContentReloaded",".overlay",function(t,r){t.button=e("#"+t.button.attr("id")),t.loadSelectedItems(t)}.bind(null,i))}function n(t){e("body").removeClass("RefBrowserOverlayOpened"),e(".refbrowser").remove()}t.prototype.loadSelectedItems=function(t){var r=t.field.find(".selected_items");r.find("ul").is(":empty")&&e(r).find('[type="hidden"]').each(function(n,i){var a={},s=e(i);if(""!==s.val()){const o=e("body").data("portal-url");a.path=s.val(),a.title=s.data("title")+" ("+a.path.substring(o.replace(window.location.origin,"").length)+")",a.selectable=!0,a.traversable=!1,a.addclass="",a.tag="span",a.selected='checked="checked"',a.type=t.sel_type,a.name=t.name,a.uid=s.data("uid"),a.checkbox=t.checkbox_template(a),s.remove();var l=e(t.list_template(a));t.AddremoveLink(l),e(r).find("ul").append(l)}})},t.prototype.openOverlay=function(t,n){n.stopPropagation(),n.preventDefault(),e("body").addClass("RefBrowserOverlayOpened"),t.list_template=r.compile(e("#listing-template",t.field).html()),t.checkbox_template=r.compile(e("#checkbox-template",t.field).html()),e(".sortable").sortable();var i=e(n.currentTarget),a=i.closest(".referencewidget");t.container=a,t.sel_type=a.data("type"),t.field_id=t.field.attr("id"),t.isInTinyMce?t.widget_url=a.data("url"):t.widget_url=a.data("url")+"/++widget++"+t.name;var s=r.compile(e("#refbrowser-template",t.field).html());e("body").append(s(t.translations)),t.build_pathbar(t,""),t.get_data(t,"");var o=e(".refbrowser");o.on("click",".path a",t.jump_to.bind(null,t)),o.on("click",".search button",t.search.bind(null,t)),o.on("keypress",".search input",function(e){13==e.which&&t.search(t,e)}),o.on("change",".listing input.ref-checkbox",t.checkbox_flipped.bind(null,t)),o.on("change","select",t.reload.bind(null,t)),o.on("click","button.cancel",function(){e(".refbrowser").remove()}),o.on("click",".ref_list_entry a",t.switch_level.bind(null,t)),o.on("click",".listing input.ref-checkbox",function(t){t.stopPropagation()}),o.on("click",".refbrowser_batching a",t.change_page.bind(null,t))},t.prototype.search=function(t,r){r.stopPropagation(),r.preventDefault();var n=e(".refbrowser .searchField").val(),i=e(".refbrowser #searchCurrentPath:checked").length;0!==n.length&&(t.term=n,t.search_current_path=i,t.page=1,t.search_results(t))},t.prototype.search_results=function(t){var r={term:t.term,search_current_path:t.search_current_path,request_path:t.request_path,page:t.page,sort_on:e('.refbrowser select[name="sort_on"]').val(),sort_order:e('.refbrowser select[name="sort_order"]').val()};e.post(t.widget_url+"/search_for_refs",r,function(t,r){e(".refbrowser .refbrowser_batching").remove(),t.build_list(t,r),t.request_data=r,e(".refbrowser .listing").addClass("search_result")}.bind(null,t))},t.prototype.build_pathbar=function(t,n){e.post(t.widget_url+"/generate_pathbar",{origin:n},function(n){var i=r.compile(e("#node-template",t.field).html()),a="";n.forEach(function(t){a+=i(t)}),e(".path").html(a)})},t.prototype.switch_level=function(t,r){r.preventDefault(),r.stopPropagation();var n=e(r.currentTarget).parent();if(!1!==n.data("traversable")){var i=n.data("path");t.request_path=i,t.build_pathbar(t,i),t.get_data(t,i),e(".refbrowser .searchField").val("")}},t.prototype.reload=function(t,r){r.preventDefault(),r.stopPropagation(),0===e(".refbrowser .searchField").val().length?t.get_data(t,t.request_path,t.page):t.search(t,r)},t.prototype.change_page=function(t,r){var n=e(r.currentTarget);n.hasClass("next")?t.page++:n.hasClass("previous")?t.page--:t.page=parseInt(n.text()),!0===n.closest(".formcontrols").siblings(".listing").hasClass("search_result")?t.search_results(t,t.page):t.get_data(t,t.request_path,t.page)},t.prototype.rebuild_listing=function(t,r){e(".refbrowser .listing ul").empty();var n="";for(var i in r)if(r.hasOwnProperty(i)){var a=r[i];a.selected="",a.extras="",a.tag="span",a.name=t.name;var s=e('.referencewidget .selected_items li[data-path="'+a.path+'"]',t.field);s.find("input:checked").length>0&&(a.selected='checked="checked"'),a.addclass="",a.traversable&&(a.addclass="traversable",a.tag="a",a.extras="href='#'"),a.selectable&&(a.type=t.sel_type,a.checkbox=t.checkbox_template(a)),n+=t.list_template(a)}e(".listing ul").append(n)},t.prototype.build_list=function(t,n){e(".refbrowser .refbrowser_batching").remove();var i=r.compile(e("#batch_template",t.field).html());e(".refbrowser .batchingcontainer").append(i(n)),e(".refbrowser .batchingcontainer .previous").html("&laquo;"),e(".refbrowser .batchingcontainer .next").html("&raquo;");var a=r.compile(e("#sort-template",t.field).html());e(".refbrowser .sorter").html(a({sortOnOptions:n.sortOnOptions,sortOrderOptions:n.sortOrderOptions,label_sort_by:t.translations.label_sort_by})),t.rebuild_listing(t,n.items);var s=e(".refbrowser .pathbar").outerHeight(!0);e(".refbrowser .listing").css({top:s+"px"})},t.prototype.get_data=function(t,r,n){void 0===n&&(n=1);var i={start:r,page:n,sort_on:e('.refbrowser select[name="sort_on"]').val(),sort_order:e('.refbrowser select[name="sort_order"]').val()};e.post(t.widget_url+"/get_reference_data",i,function(t,e){t.request_data=e,t.build_list(t,t.request_data)}.bind(null,t))},t.prototype.jump_to=function(t,r){r.stopPropagation(),r.preventDefault();var n=e(r.currentTarget),i=e(n).data("path");e(n).data("clickable")&&(t.request_path=i,t.build_pathbar(t,i),t.get_data(t,i),e(".refbrowser .searchField").val(""))},t.prototype.resize=function(){e(".refbrowser .pathbar").each(function(t){var r=e(this).outerHeight(!0);e(this).siblings(".listing").css({top:r+"px"})})},t.prototype.AddremoveLink=function(t){if(1===e('[type="radio"]',t).length){var r=e('<a href="#" class="removeItem">X</a>');r.on("click",function(t){t.preventDefault(),t.stopPropagation(),e(this).parent().remove()}),e(t).prepend(r)}return t},t.prototype.checkbox_flipped=function(t,r){r.stopPropagation(),r.preventDefault();var i=r.currentTarget;if(!0===i.checked){var a=e(r.currentTarget.parentNode).clone();e(a).find("input").attr("name",t.name),e(a).find("a").on("click",function(t){t.preventDefault()});var s=e.trim(a.find("span").text());s=s+" ("+a.data("path")+")",a.find("span").text(s),t.AddremoveLink(a),e(".selected_items ul",t.container).append(a),"radio"===e(i).attr("type")&&n()}else{var o="#"+t.field_id+' .referencewidget .selected_items li[data-path="'+e(r.currentTarget.parentNode).data("path")+'"]';e(o).remove()}},e(window).on("load",function(r){var n=e(".referencewidget button");0!==n.length&&n.each(function(e,r){new t(r)})}),e(document).on("onLoad",".overlay",function(r){var n=e(".referencewidget button",e(this));0!==n.length&&n.each(function(e,r){new t(r)})}),window.initRefBrowser=t})}(window,window.jQuery,window.Handlebars),define("refwidget",function(){}),define("text",["module"],function(t){"use strict";function e(t,e){return void 0===t||""===t?e:t}function r(t,r,n,i){if(r===i)return!0;if(t===n){if("http"===t)return e(r,"80")===e(i,"80");if("https"===t)return e(r,"443")===e(i,"443")}return!1}var n,i,a,s,o,l=["Msxml2.XMLHTTP","Microsoft.XMLHTTP","Msxml2.XMLHTTP.4.0"],c=/^\s*<\?xml(\s)+version=[\'\"](\d)*.(\d)*[\'\"](\s)*\?>/im,p=/<body[^>]*>\s*([\s\S]+)\s*<\/body>/im,h="undefined"!=typeof location&&location.href,u=h&&location.protocol&&location.protocol.replace(/\:/,""),d=h&&location.hostname,f=h&&(location.port||void 0),m={},g=t.config&&t.config()||{};return n={version:"2.0.16",strip:function(t){if(t){t=t.replace(c,"");var e=t.match(p);e&&(t=e[1])}else t="";return t},jsEscape:function(t){return t.replace(/(['\\])/g,"\\$1").replace(/[\f]/g,"\\f").replace(/[\b]/g,"\\b").replace(/[\n]/g,"\\n").replace(/[\t]/g,"\\t").replace(/[\r]/g,"\\r").replace(/[\u2028]/g,"\\u2028").replace(/[\u2029]/g,"\\u2029")},createXhr:g.createXhr||function(){var t,e,r;if("undefined"!=typeof XMLHttpRequest)return new XMLHttpRequest;if("undefined"!=typeof ActiveXObject)for(e=0;e<3;e+=1){r=l[e];try{t=new ActiveXObject(r)}catch(t){}if(t){l=[r];break}}return t},parseName:function(t){var e,r,n,i=!1,a=t.lastIndexOf("."),s=0===t.indexOf("./")||0===t.indexOf("../");return-1!==a&&(!s||a>1)?(e=t.substring(0,a),r=t.substring(a+1)):e=t,n=r||e,a=n.indexOf("!"),-1!==a&&(i="strip"===n.substring(a+1),n=n.substring(0,a),r?r=n:e=n),{moduleName:e,ext:r,strip:i}},xdRegExp:/^((\w+)\:)?\/\/([^\/\\]+)/,useXhr:function(t,e,i,a){var s,o,l,c=n.xdRegExp.exec(t);return!c||(s=c[2],o=c[3],o=o.split(":"),l=o[1],o=o[0],(!s||s===e)&&(!o||o.toLowerCase()===i.toLowerCase())&&(!l&&!o||r(s,l,e,a)))},finishLoad:function(t,e,r,i){r=e?n.strip(r):r,g.isBuild&&(m[t]=r),i(r)},load:function(t,e,r,i){if(i&&i.isBuild&&!i.inlineText)return void r();g.isBuild=i&&i.isBuild;var a=n.parseName(t),s=a.moduleName+(a.ext?"."+a.ext:""),o=e.toUrl(s),l=g.useXhr||n.useXhr;if(0===o.indexOf("empty:"))return void r();!h||l(o,u,d,f)?n.get(o,function(e){n.finishLoad(t,a.strip,e,r)},function(t){r.error&&r.error(t)}):e([s],function(t){n.finishLoad(a.moduleName+"."+a.ext,a.strip,t,r)},function(t){r.error&&r.error(t)})},write:function(t,e,r,i){if(m.hasOwnProperty(e)){var a=n.jsEscape(m[e]);r.asModule(t+"!"+e,"define(function () { return '"+a+"';});\n")}},writeFile:function(t,e,r,i,a){var s=n.parseName(e),o=s.ext?"."+s.ext:"",l=s.moduleName+o,c=r.toUrl(s.moduleName+o)+".js";n.load(l,r,function(e){var r=function(t){return i(c,t)};r.asModule=function(t,e){return i.asModule(t,c,e)},n.write(t,l,r,a)},a)}},"node"===g.env||!g.env&&"undefined"!=typeof process&&process.versions&&process.versions.node&&!process.versions["node-webkit"]&&!process.versions["atom-shell"]?(i=require.nodeRequire("fs"),n.get=function(t,e,r){try{var n=i.readFileSync(t,"utf8");"\ufeff"===n[0]&&(n=n.substring(1)),e(n)}catch(t){r&&r(t)}}):"xhr"===g.env||!g.env&&n.createXhr()?n.get=function(t,e,r,i){var a,s=n.createXhr();if(s.open("GET",t,!0),i)for(a in i)i.hasOwnProperty(a)&&s.setRequestHeader(a.toLowerCase(),i[a]);g.onXhr&&g.onXhr(s,t),s.onreadystatechange=function(n){var i,a;4===s.readyState&&(i=s.status||0,i>399&&i<600?(a=new Error(t+" HTTP status: "+i),a.xhr=s,r&&r(a)):e(s.responseText),g.onXhrComplete&&g.onXhrComplete(s,t))},s.send(null)}:"rhino"===g.env||!g.env&&"undefined"!=typeof Packages&&"undefined"!=typeof java?n.get=function(t,e){var r,n,i=new java.io.File(t),a=java.lang.System.getProperty("line.separator"),s=new java.io.BufferedReader(new java.io.InputStreamReader(new java.io.FileInputStream(i),"utf-8")),o="";try{for(r=new java.lang.StringBuffer,n=s.readLine(),n&&n.length()&&65279===n.charAt(0)&&(n=n.substring(1)),null!==n&&r.append(n);null!==(n=s.readLine());)r.append(a),r.append(n);o=String(r.toString())}finally{s.close()}e(o)}:("xpconnect"===g.env||!g.env&&"undefined"!=typeof Components&&Components.classes&&Components.interfaces)&&(a=Components.classes,s=Components.interfaces,Components.utils.import("resource://gre/modules/FileUtils.jsm"),o="@mozilla.org/windows-registry-key;1"in a,n.get=function(t,e){var r,n,i,l={};o&&(t=t.replace(/\//g,"\\")),i=new FileUtils.File(t);try{r=a["@mozilla.org/network/file-input-stream;1"].createInstance(s.nsIFileInputStream),r.init(i,1,0,!1),n=a["@mozilla.org/intl/converter-input-stream;1"].createInstance(s.nsIConverterInputStream),n.init(r,"utf-8",r.available(),s.nsIConverterInputStream.DEFAULT_REPLACEMENT_CHARACTER),n.readString(r.available(),l),n.close(),r.close(),e(l.value)}catch(t){throw new Error((i&&i.path||"")+": "+t)}}),n}),define("text!link.xml",[],function(){return'<div>\n  <div class="linkModal">\n    <h1><%- insertHeading %></h1>\n    <% if(upload){ %>\n    <p class="info"><%- insertLinkHelp %></p>\n    <% } %>\n\n    <div class="linkTypes pat-autotoc autotabs"\n         data-pat-autotoc="section:fieldset;levels:legend;IDPrefix:tinymce-autotoc-">\n\n      <fieldset class="linkType internal" data-linkType="internal">\n        <legend id="tinylink-internal"><%- internal %></legend>\n        <div>\n          <div class="form-group main">\n            <div class="referencewidget referencewidgetField"\n                data-url="<%- options.base_url %>"\n                data-path="<%- options.relatedItems.basePath %>"\n                data-fieldname="internal"\n                data-type="radio"\n                data-trans=\'{"close": "Schliessen", "search": "Suchen", "label_sort_by": "Sortieren nach", "search_current_path": "Suche nur im aktuellen Pfad"}\'>\n\n                <div class="selected_items">\n                    <ul></ul>\n                </div>\n                <input type="hidden" name="internal" />\n                <button type="button"\n                        id="browse-internal-link"\n                        name="browse-internal">Durchsuchen</button>\n\n\n            <script id="refbrowser-template" type="text/x-handlebars-template">\n                <div class="refbrowser overlay">\n                        <div class="pathbar">\n                            <div class="path"></div>\n                            <div class="search">\n                                <input class="searchField" name="refbrowser-search" type="text" />\n                                <button class="searchButton" name="refbrowser-search" type="submit">{{search}}</button>\n                                <label for="searchCurrentPath" class="searchCurrentPathLabel">\n                                    {{search_current_path}}\n                                    <input type="checkbox" name="refbrowser-search-current-path" id="searchCurrentPath">\n                                </label>\n                            </div>\n                        <div class="sorter"></div>\n                        </div>\n\n                        <div class="listing">\n                            <ul></ul>\n                        </div>\n                        <div class="formcontrols">\n                            <div class="closecontainer">\n                             <button type="cancel" class="cancel">{{close}}</button>\n                            </div>\n                            <div class="batchingcontainer" />\n                        </div>\n                </div>\n            <\/script>\n\n            <script id="node-template" type="text/x-handlebars-template">\n                <a data-path="{{path}}" href="#" data-clickable={{clickable}}>{{title}}</a>\n            <\/script>\n\n            <script id="listing-template" type="text/x-handlebars-template">\n                <li class="ref_list_entry {{addclass}}" data-traversable="{{traversable}}" data-path="{{path}}" data-id="{{id}}">\n                    {{{checkbox}}}\n                    <{{{tag}}} class="{{content-type}}" {{{extras}}}> {{title}} </{{{tag}}}>\n                </li>\n            <\/script>\n\n            <script id="checkbox-template" type="text/x-handlebars-template">\n                    <input type="{{type}}" class="ref-checkbox" {{{selected}}} name="{{name}}" value="{{path}}" data-uid="{{uid}}"/>\n            <\/script>\n\n            <script id="batch_template" type="text/x-handlebars-template">\n                    <div class="refbrowser_batching">\n                        {{{batching}}}\n                    </div>\n            <\/script>\n\n            <script id="sort-template" type="text/x-handlebars-template">\n                {{label_sort_by}}\n                <select name="sort_on">\n                    {{#each sortOnOptions}}\n                        <option value="{{this.value}}" {{#if this.selected}}SELECTED="SELECTED"{{/if}}>\n                            {{this.title}}\n                        </option>\n                    {{/each}}\n                </select>\n                <select name="sort_order">\n                    {{#each sortOrderOptions}}\n                        <option value="{{this.value}}" {{#if this.selected}}SELECTED="SELECTED"{{/if}}>\n                            {{this.title}}\n                        </option>\n                    {{/each}}\n                </select>\n\n            <\/script>\n\n\n\n            </div>\n          </div>\n        </div>\n      </fieldset>\n\n      <% if(upload){ %>\n      <fieldset class="linkType upload" data-linkType="upload">\n        <legend id="tinylink-upload"><%- uploadText %></legend>\n        <div class="uploadify-me"></div>\n      </fieldset>\n      <% } %>\n\n      <fieldset class="linkType external" data-linkType="external">\n        <legend id="tinylink-external"><%- external %></legend>\n        <div class="form-group main">\n          <label for="external"><%- externalText %></label>\n          <input type="text" name="external" />\n        </div>\n      </fieldset>\n\n      <fieldset class="linkType email" data-linkType="email">\n        <legend id="tinylink-email"><%- emailText %></legend>\n        <div class="form-inline">\n          <div class="form-group main">\n            <label><%- emailText %></label>\n            <input type="text" name="email" />\n          </div>\n          <div class="form-group">\n            <label><%- subjectText %></label>\n            <input type="text" name="subject" />\n          </div>\n        </div>\n      </fieldset>\n\n      <fieldset class="linkType anchor" data-linkType="anchor">\n        <legend id="tinylink-anchor"><%- anchor %></legend>\n        <div>\n          <div class="form-group main">\n            <label><%- anchorLabel %></label>\n            <div class="input-wrapper">\n              <select name="anchor" class="pat-select2" data-pat-select2="width:500px" />\n            </div>\n          </div>\n        </div>\n      </fieldset>\n\n    </div>\x3c!-- / tabs --\x3e\n\n    <div class="common-controls">\n      <div class="form-group">\n        <label><%- target %></label>\n        <select name="target">\n          <% _.each(targetList, function(target){ %>\n            <option value="<%- target.value %>"><%- target.text %></option>\n          <% }); %>\n        </select>\n      </div>\n      <div class="form-group">\n        <label><%- titleText %></label>\n        <input type="text" name="title" />\n      </div>\n    </div>\n\n    <input type="submit" class="plone-btn" name="cancel" value="<%- cancelBtn %>" />\n    <input type="submit" class="plone-btn plone-btn-primary context" name="insert" value="<%- insertBtn %>" />\n  </div>\n</div>\n'}),define("refwidget-tinymce",["jquery","underscore","pat-registry","tinymce","mockup-patterns-tinymce-url/js/links","refwidget","text!link.xml","mockup-patterns-tinymce"],function(t,e,r,n,i,a,s,o){"use strict";var l=r.patterns.linktype;delete r.patterns.internallinktype,delete t.fn.internallinktype;var c=l.extend({name:"internallinktype",trigger:".pat-internallinktype-dummy",init:function(){l.prototype.init.call(this),this.refwidget=new initRefBrowser(this.getEl().next()[0])},getEl:function(){var t=this.$el.find('input[name="internal"]:checked');return t.length?t:this.$el.find('input[name="internal"]:first')},value:function(){return this.getEl().data("uid")},toUrl:function(){var t=this.value();return t?this.tinypattern.generateUrl({UID:t}):null},load:function(t){var e=this.tiny.dom.getAttrib(t,"data-val");e&&this.set(e)},set:function(e){var r=this.refwidget,n=r.field.find(".selected_items"),i=t('<input type="hidden" name="internal" />'),a=r.field.data("url")+"/search_for_refs";t.post(a,{uid:e},function(t){""!==t.items.length&&(i.data("title",t.items[0].title),i.data("uid",t.items[0].uid),i.val(t.items[0].path),i.appendTo(n),r.loadSelectedItems(r))})},attributes:function(){var t=this.value();return t?{"data-val":t}:{}}});delete r.patterns.linkmodal,delete t.fn.linkmodal;var p=i.extend({defaults:{anchorSelector:"h1,h2,h3",linkTypes:[],initialLinkType:"internal",text:{insertHeading:"Insert Link"},linkTypeClassMapping:{internal:c,upload:r.patterns.uploadlinktype,external:r.patterns.externallinktype,email:r.patterns.emaillinktype,anchor:r.patterns.anchorlinktype,image:r.patterns.imagelinktype,uploadImage:r.patterns.uploadlinktype,externalImage:l}},linkTypeTemplateMapping:{internal:s,upload:s,external:s,email:s,anchor:s,image:i.prototype.linkTypeTemplateMapping.image,uploadImage:i.prototype.linkTypeTemplateMapping.image,externalImage:i.prototype.linkTypeTemplateMapping.image}});return o.prototype.addLinkClicked=function(){var e=this;if(null===e.linkModal){var r=t("<div/>").insertAfter(e.$el),n=["internal","upload","external","email","anchor"];e.options.upload||n.splice(1,1),e.linkModal=new p(r,t.extend(!0,{},e.options,{tinypattern:e,linkTypes:n})),e.linkModal.show()}else e.linkModal.reinitialize(),e.linkModal.show()},p}),require(["handlebars","refwidget","refwidget-tinymce"],function(t,e,r){}),define("main",function(){});
