--- ftw/referencewidget/resources/refwidget.css
@@ -1,159 +1,0 @@
-div.refbrowser .listing li {
-    display:block;
-    min-height: 1.4em;
-    word-break: break-all;
-}
-
-div.refbrowser .pathbar{
-    padding: 1em 1em 1em 1em;
-    width:100%;
-    min-height:80px;
-    zoom: 1;
-    border-bottom: 1px #ddd solid;
-    position: relative;
-}
-
-div.refbrowser .path {
-    min-width: 65%;
-    width:auto;
-    float:left;
-    max-width: 80%;
-}
-
-.refbrowser .sorter {
-  position: absolute;
-  bottom: 0;
-  margin-bottom: 0.5em;
-}
-
-div.refbrowser .pathbar a:not(:last-child)::after{
-    content: "\2192";
-    display:inline-block;
-    padding-left: 3px;
-    padding-right: 3px;
-
-
-}
-
-div.refbrowser .pathbar a:hover::after{
-    color: initial;
-    text-decoration: none !important;
-
-}
-
-  div.refbrowser .pathbar:before, div.refbrowser .pathbar:after {
-    content: "\0020";
-    display: block;
-    height: 0;
-    visibility: hidden;
-  }
-  div.refbrowser .pathbar:after { clear: both; }
-
-
-div.refbrowser .search{
-    float: right;
-}
-div.refbrowser .search input{
-    min-width: 180px;
-}
-div.refbrowser .search input#searchCurrentPath{
-    min-width: auto;
-}
-div.refbrowser .search .searchCurrentPathLabel{
-    display: block;
-}
-
-div.refbrowser .search button{
-    float:right;
-}
-div.refbrowser .listing {
-    overflow-y: auto;
-    position: absolute;
-    bottom: 4em;
-    left: 1.2em;
-    right: 1.2em;
-    margin-top: .7em;
-
-}
-div.refbrowser .listing ul {
-    margin: 0;
-    padding: 0;
-
-}
-
-div.refbrowser {
-    position: fixed;
-    left: 15%;
-    right: 15%;
-    top: 10%;
-    background: #ffffff;
-    z-index: 2147483647 !important;
-    bottom: 5%;
-}
-
-div.refbrowser .shadow{
-    max-height: 64vh;
-    overflow-y: auto;
-}
-div.refbrowser .formcontrols {
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    background:#f5f5f5;
-    padding: .5em 1.2em;
-    height:4em;
-}
-
-div.refbrowser .formcontrols button {
-    margin:0;
-}
-
-div.refbrowser .closecontainer {
-    width: 20%;
-    float:left;
-}
-
-div.refbrowser .batchingcontainer {
-    width:80%;
-    float:left;
-}
-
-div.refbrowser .formcontrols .pagination.listingBar {
-    margin: 0;
-}
-div.refbrowser .formcontrols .pagination.listingBar a {
-    margin-bottom: 0;
-}
-
-div.refbrowser.overlay {
-    display:block;
-}
-
-.selected_items ul {
-  margin-left: 0;
-  padding-left: 0;
-  list-style-type: none;
-}
-
-.ref_list_entry::before {
-  display: inline-block;
-  width: 16px;
-  content: "";
-}
-
-.ref_list_entry.traversable::before {
-  font-family: 'FontAwesome';
-  content: "\f196";
-  margin: 0px 5px 0 -5px;
-}
-
-li.ref_list_entry.traversable:hover,
-li.ref_list_entry.traversable:hover a {
-  color: #75ad0a;
-}
-
-.ref_list_entry a.removeItem {
-  display: inline-block;
-  margin: 0 .5em;
-}
