+++ ftw/referencewidget/templates/referencewidget_input.pt
@@ -2,13 +2,6 @@
       xmlns:i18n="http://xml.zope.org/namespaces/i18n"
       i18n:domain="ftw.referencewidget">
 
-<div class="referencewidget referencewidgetField"
-     tal:attributes="name view/name;
-                     data-url view/form_url;
-                     data-path view/current_path;
-                     data-fieldname view/name;
-                     data-type view/is_list;
-                     data-trans view/translations">
 <div class="selected_items">
     <ul class="sortable"></ul>
     <!-- Rendered from the backend. This is gonna be converted to a list placed in the ul above by javascript-->
@@ -20,13 +13,12 @@
 </div>
 
 <input type="hidden" tal:attributes="name view/name" />
-<button type="button" tal:attributes="id string:browse-${view/id};
-                                      name string:browse-${view/name};
-                                      class string:context browse-${view/klass};"
-                        i18n:translate="">Browse</button>
 
-
-<tal:HANDLEBARS replace="structure view/handlebars_html" />
-</div>
+<div class="reference-widget referencewidget referencewidgetField"
+     tal:attributes="name view/name;
+                     data-starturl view/get_start_url;
+                     data-fieldname view/name;
+                     data-type view/is_list;
+                     data-translations view/translations"/>
 
 </tal:html>
