+++ package.json
@@ -1,19 +1,36 @@
 {
-  "name": "ftw.referencewidget",
-  "version": "1.0.0",
-  "repository": "git@github.com:4teamwork/ftw.referencewidget.git",
-  "author": "4teamwork AG <info@4teamwork.ch>",
-  "license": "GPL2",
-  "scripts": {
-    "build": "r.js -o build.js && r.js -o buildcss.js",
-    "build-js": "r.js -o build.js",
-    "build-css": "r.js -o buildcss.js",
-    "watch": "chokidar \"./ftw/referencewidget/resources/(*.js|*.css)\" -i \"**/*-compiled.(js|css)\" -c \"if [[ '{path}' = *.js ]]; then r.js -o build.js; else r.js -o buildcss.js; fi;\""
-  },
-  "private": true,
-  "devDependencies": {
-    "chokidar-cli": "^2.1.0",
-    "requirejs": "^2.3.6",
-    "text": "github:requirejs/text"
-  }
+    "private": true,
+    "name": "simplelayout",
+    "description": null,
+    "scripts": {
+        "dev": "webpack --watch --progress",
+        "build": "webpack --env prod --progress"
+    },
+    "dependencies": {
+        "@patternslib/patternslib": "8.1.0-alpha.1",
+        "@plone/mockup": "^5.0.0-alpha.7",
+        "axios": "^0.27.2",
+        "qs": "^6.10.5",
+        "vue": "^3.2.36",
+        "vue-axios": "^3.4.1"
+    },
+    "devDependencies": {
+        "@rushstack/eslint-patch": "^1.1.0",
+        "@vue/compiler-sfc": "^3.2.37",
+        "@vue/eslint-config-prettier": "^7.0.0",
+        "copy-webpack-plugin": "^11.0.0",
+        "css-loader": "^6.7.1",
+        "css-minimizer-webpack-plugin": "^4.0.0",
+        "eslint": "^8.5.0",
+        "eslint-plugin-vue": "^8.2.0",
+        "file-loader": "^6.2.0",
+        "mini-css-extract-plugin": "^2.6.0",
+        "prettier": "^2.5.1",
+        "sass": "^1.52.2",
+        "sass-loader": "^13.0.0",
+        "url-loader": "^4.1.1",
+        "vue-loader": "^17.0.0",
+        "webpack": "^5.73.0",
+        "webpack-cli": "^4.9.2"
+    }
 }
