+++ sources.cfg
@@ -6,8 +6,10 @@ auto-checkout =
     ftw.testing
     ftw.testbrowser
     collective.z3cform.datagridfield
+    plone.staticresources
 
 [branches]
 ftw.testing = plone6
 ftw.testbrowser = mle-plone6-support
 collective.z3cform.datagridfield = petschki-autoform-dictrow-fix
+
